refactor(contract): split lib.rs into domain modules
Split the monolithic `predifi-contract/src/lib.rs` (6,673 lines / 260KB)
into six focused domain modules, following the module conventions already
established in the crate (constants.rs, events.rs, payouts.rs, ...).

  lib.rs        6673 -> 2030 lines
  admin.rs      1239 lines   37 entrypoints
  pool.rs       1584 lines   20 entrypoints
  prediction.rs  930 lines    5 entrypoints
  oracle.rs      667 lines    8 entrypoints + OracleCallback impl
  referral.rs    171 lines    7 entrypoints
  treasury.rs    144 lines    3 entrypoints

This is a pure code move. No function body, signature, storage key, event
or error was modified.

Structure: each domain module carries its own `#[contractimpl] impl
PredifiContract` block. lib.rs retains the `#[contract] pub struct`, all
shared types, constants, events, the OracleCallback trait, and the 33
cross-domain private helpers (require_*, extend_*, get_config,
validate_*, ...), which stay visible to every submodule because they live
in the crate-root impl block.

The only edits beyond relocation are the `use` statements each new module
needs, including `PredifiContractArgs` / `PredifiContractClient`, which
`#[contractimpl]` requires in scope when it expands outside the module
that hosts `#[contract]`.

Verified: the contract interface dumped from the release WASM is
unchanged - 81 exported functions, identical signature set, identical
spec (only emission order differs).

Refs #1403
