+++ contract/contracts/predifi-contract/src/lib.rs
@@ -1,16 +1,21 @@
 #![no_std]
 #![allow(clippy::too_many_arguments)]
 
+mod admin;
 mod benchmark_test;
 #[cfg(test)]
 mod boundary_edge_case_tests;
 mod constants;
 mod gas_opt;
-mod payouts;
+mod oracle;
 #[cfg(test)]
 mod payout_proptests;
+mod payouts;
+mod pool;
+mod prediction;
 mod price_feed;
 mod price_feed_simple;
+mod referral;
 mod safe_math;
 #[cfg(test)]
 mod safe_math_examples;
@@ -20,9 +25,10 @@ mod storage_test;
 mod stress_test;
 #[cfg(test)]
 mod test_utils;
+mod treasury;
 
 use soroban_sdk::{
-    contract, contracterror, contractevent, contractimpl, contracttype, log, symbol_short, token,
+    contract, contracterror, contractevent, contractimpl, contracttype, symbol_short, token,
     Address, BytesN, Env, IntoVal, String, Symbol, SymbolStr, TryFromVal, Vec,
 };
 
@@ -1982,4236 +1988,9 @@ impl PredifiContract {
         Self::extend_persistent(env, &ctr_key);
     }
 
-    // ── Public interface ──────────────────────────────────────────────────────
-
-    /// Initialize the contract. Idempotent — safe to call multiple times.
-    pub fn init(
-        env: Env,
-        access_control: Address,
-        treasury: Address,
-        fee_bps: u32,
-        resolution_delay: u64,
-        min_pool_duration: u64,
-        max_predictions_per_user: u32,
-    ) {
-        if env.storage().instance().has(&DataKey::Config) {
-            soroban_sdk::panic_with_error!(&env, PredifiError::AlreadyInitializedOrConfigNotSet);
-        }
-
-        // Enforce the same 30-day cap on resolution_delay that
-        // set_resolution_delay enforces, so the contract cannot be
-        // initialised with an unbounded delay.
-        if resolution_delay > MAX_RESOLUTION_DELAY {
-            soroban_sdk::panic_with_error!(&env, PredifiError::InvalidData);
-        }
-
-        // Validate fee_bps on init — consistent with set_fee_bps (INV-6)
-        if !Self::is_valid_fee_bps(fee_bps) {
-            soroban_sdk::panic_with_error!(&env, PredifiError::InvalidFeeBps);
-        }
-
-        let config = Config {
-            fee_bps,
-            treasury: treasury.clone(),
-            access_control: access_control.clone(),
-            resolution_delay,
-            min_pool_duration,
-            min_stake: DEFAULT_GLOBAL_MIN_STAKE,
-            max_predictions_per_user,
-            prediction_cooldown_seconds: DEFAULT_PREDICTION_COOLDOWN_SECONDS,
-            referral_bps: 5000,              // default 50%
-            claim_window_seconds: 2_592_000, // default 30 days
-        };
-        env.storage().instance().set(&DataKey::Config, &config);
-        env.storage().instance().set(&DataKey::PoolIdCtr, &0u64);
-        env.storage()
-            .instance()
-            .set(&DataKey::Version, &CONTRACT_VERSION);
-        Self::extend_instance(&env);
-
-        InitEvent {
-            access_control,
-            treasury,
-            fee_bps,
-            resolution_delay,
-            min_pool_duration,
-            max_predictions_per_user,
-        }
-        .publish(&env);
-    }
-
-    /// Pause the contract.
-    ///
-    /// # Authorization
-    /// Requires authentication from a caller with the Admin role.
-    ///
-    /// # Effects
-    /// - Marks the contract as paused.
-    /// - Emits `ContractPausedAlertEvent`.
-    /// - Emits `PauseEvent`.
-    ///
-    /// While paused, administrative checks continue to work, but
-    /// state-changing operations guarded by the pause flag are rejected.
-    pub fn pause(env: Env, admin: Address) {
-        admin.require_auth();
-        if Self::require_admin_role(&env, &admin, "pause").is_err() {
-            panic!("Unauthorized: missing required role");
-        }
-        env.storage().instance().set(&DataKey::Paused, &true);
-        Self::extend_instance(&env);
-
-        ContractPausedAlertEvent {
-            admin: admin.clone(),
-            timestamp: env.ledger().timestamp(),
-        }
-        .publish(&env);
-
-        PauseEvent { admin }.publish(&env);
-    }
-
-    /// Resume normal contract operation.
-    ///
-    /// # Authorization
-    /// Requires authentication from a caller with the Admin role.
-    ///
-    /// # Effects
-    /// - Clears the paused state.
-    /// - Emits `UnpauseEvent`.
-    pub fn unpause(env: Env, admin: Address) {
-        admin.require_auth();
-        if Self::require_admin_role(&env, &admin, "unpause").is_err() {
-            panic!("Unauthorized: missing required role");
-        }
-        env.storage().instance().set(&DataKey::Paused, &false);
-        Self::extend_instance(&env);
-
-        UnpauseEvent { admin }.publish(&env);
-    }
-
-    /// Check if the contract is paused.
-    ///
-    /// This is a public query function that allows third-party integrations
-    /// to check the pause state without sending a transaction.
-    ///
-    /// # Returns
-    /// `true` if the contract is paused, `false` otherwise.
-    pub fn is_contract_paused(env: Env) -> bool {
-        Self::is_paused(&env)
-    }
-
-    /// Return the contract version stored in instance storage.
-    /// Returns 0 if the contract was deployed before version tracking was added.
-    pub fn get_version(env: Env) -> u32 {
-        env.storage()
-            .instance()
-            .get(&DataKey::Version)
-            .unwrap_or(0u32)
-    }
-
-    /// Return the contract version as a semantic version string.
-    ///
-    /// This getter provides the human-readable version number in the format "X_Y_Z"
-    /// (e.g., "0_0_0"). The version string matches the version specified in Cargo.toml
-    /// but uses underscores instead of dots since Symbols don't allow dots.
-    ///
-    /// # Returns
-    /// A `Symbol` containing the current contract version string.
-    ///
-    /// # Example
-    /// ```ignore
-    /// let version = contract.get_version_string(&env);
-    /// assert_eq!(version, Symbol::new(&env, "0_0_0"));
-    /// ```
-    pub fn get_version_string(env: Env) -> Symbol {
-        Symbol::new(&env, "0_0_0")
-    }
-
-    /// Queue a fee change proposal subject to a [`FEE_CHANGE_TIMELOCK_SECONDS`]-second
-    /// delay. The new fee does **not** take effect immediately; the admin must call
-    /// [`Self::apply_fee_bps`] once the delay has elapsed.
-    ///
-    /// # Errors
-    /// - [`PredifiError::ContractPaused`]   – contract is paused.
-    /// - [`PredifiError::Unauthorized`]     – caller lacks Admin role (0).
-    /// - [`PredifiError::InvalidFeeBps`]    – `fee_bps > 10_000`.
-    /// - [`PredifiError::FeeChangePending`] – a proposal is already queued;
-    ///   call `cancel_fee_proposal` first.
-    ///
-    /// PRE:  admin has role 0; no pending fee proposal exists.
-    /// POST: A [`PendingFeeChange`] is stored with
-    ///       `effective_at = now + FEE_CHANGE_TIMELOCK_SECONDS`.
-    pub fn set_fee_bps(env: Env, admin: Address, fee_bps: u32) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "set_fee_bps")?;
-        if !Self::is_valid_fee_bps(fee_bps) {
-            return Err(PredifiError::InvalidFeeBps);
-        }
-        // Only one proposal may be pending at a time.
-        if env.storage().instance().has(&DataKey::PendingFeeBps) {
-            return Err(PredifiError::FeeChangePending);
-        }
-        let effective_at = env.ledger().timestamp() + FEE_CHANGE_TIMELOCK_SECONDS;
-        let pending = PendingFeeChange {
-            new_fee_bps: fee_bps,
-            effective_at,
-            proposed_by: admin.clone(),
-        };
-        env.storage()
-            .instance()
-            .set(&DataKey::PendingFeeBps, &pending);
-        Self::extend_instance(&env);
-
-        FeeChangeProposeEvent {
-            admin,
-            new_fee_bps: fee_bps,
-            effective_at,
-        }
-        .publish(&env);
-        Ok(())
-    }
-
-    /// Apply a pending fee change once the [`FEE_CHANGE_TIMELOCK_SECONDS`]-second
-    /// delay has elapsed.
-    ///
-    /// Emits a [`FeeUpdateEvent`] with the newly committed fee value.
-    ///
-    /// # Errors
-    /// - [`PredifiError::ContractPaused`]      – contract is paused.
-    /// - [`PredifiError::Unauthorized`]        – caller lacks Admin role (0).
-    /// - [`PredifiError::NoFeeChangePending`]  – no proposal is queued.
-    /// - [`PredifiError::TimelockNotExpired`]  – the delay has not yet elapsed.
-    ///
-    /// PRE:  a [`PendingFeeChange`] exists and `now >= effective_at`.
-    /// POST: `Config.fee_bps` is updated; the pending proposal is removed.
-    pub fn apply_fee_bps(env: Env, admin: Address) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "apply_fee_bps")?;
-
-        let pending: PendingFeeChange = env
-            .storage()
-            .instance()
-            .get(&DataKey::PendingFeeBps)
-            .ok_or(PredifiError::NoFeeChangePending)?;
-
-        if env.ledger().timestamp() < pending.effective_at {
-            return Err(PredifiError::TimelockNotExpired);
-        }
-
-        let mut config = Self::get_config(&env);
-        config.fee_bps = pending.new_fee_bps;
-        env.storage().instance().set(&DataKey::Config, &config);
-        env.storage().instance().remove(&DataKey::PendingFeeBps);
-        Self::extend_instance(&env);
-
-        FeeUpdateEvent {
-            admin,
-            fee_bps: pending.new_fee_bps,
-        }
-        .publish(&env);
-        Ok(())
-    }
-
-    /// Cancel a pending fee change proposal before it is applied.
-    ///
-    /// Emits a [`FeeChangeCancelEvent`]. The current `Config.fee_bps` is unchanged.
-    ///
-    /// # Errors
-    /// - [`PredifiError::ContractPaused`]     – contract is paused.
-    /// - [`PredifiError::Unauthorized`]       – caller lacks Admin role (0).
-    /// - [`PredifiError::NoFeeChangePending`] – no proposal is queued.
-    ///
-    /// PRE:  a [`PendingFeeChange`] exists.
-    /// POST: the pending proposal is removed; `Config.fee_bps` is unmodified.
-    pub fn cancel_fee_proposal(env: Env, admin: Address) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "cancel_fee_proposal")?;
-
-        if !env.storage().instance().has(&DataKey::PendingFeeBps) {
-            return Err(PredifiError::NoFeeChangePending);
-        }
-
-        env.storage().instance().remove(&DataKey::PendingFeeBps);
-        Self::extend_instance(&env);
-
-        FeeChangeCancelEvent { admin }.publish(&env);
-        Ok(())
-    }
-
-    /// Return the currently pending fee change proposal, or `None` if no proposal
-    /// is queued.
-    ///
-    /// Clients should poll this before calling [`Self::apply_fee_bps`] to confirm
-    /// a proposal exists and to read the `effective_at` timestamp.
-    pub fn get_pending_fee_change(env: Env) -> Option<PendingFeeChange> {
-        env.storage().instance().get(&DataKey::PendingFeeBps)
-    }
-
-    /// Set treasury address. Caller must have Admin role (0).
-    pub fn set_treasury(env: Env, admin: Address, treasury: Address) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "set_treasury")?;
-        let mut config = Self::get_config(&env);
-        config.treasury = treasury.clone();
-        env.storage().instance().set(&DataKey::Config, &config);
-        Self::extend_instance(&env);
-
-        TreasuryUpdateEvent { admin, treasury }.publish(&env);
-        Ok(())
-    }
-
-    /// Set maximum predictions per user. Caller must have Admin role (0).
-    /// PRE: admin has role 0
-    /// POST: Config.max_predictions_per_user >= 0 (0 = no limit)
-    pub fn set_max_predictions_per_user(
-        env: Env,
-        admin: Address,
-        limit: u32,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "set_max_predictions_per_user")?;
-        let mut config = Self::get_config(&env);
-        config.max_predictions_per_user = limit;
-        env.storage().instance().set(&DataKey::Config, &config);
-        Self::extend_instance(&env);
-
-        MaxPredictionsUpdateEvent { admin, limit }.publish(&env);
-        Ok(())
-    }
-
-    /// Set the cooldown in seconds between consecutive predictions from the same address.
-    pub fn set_prediction_cooldown(
-        env: Env,
-        admin: Address,
-        cooldown_seconds: u64,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "set_prediction_cooldown")?;
-
-        let mut config = Self::get_config(&env);
-        config.prediction_cooldown_seconds = cooldown_seconds;
-        env.storage().instance().set(&DataKey::Config, &config);
-        Self::extend_instance(&env);
-
-        PredictionCooldownUpdateEvent {
-            admin,
-            cooldown_seconds,
-        }
-        .publish(&env);
-        Ok(())
-    }
-
-    /// Set resolution delay in seconds. Caller must have Admin role (0).
-    pub fn set_resolution_delay(env: Env, admin: Address, delay: u64) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "set_resolution_delay")?;
-        if delay > MAX_RESOLUTION_DELAY {
-            return Err(PredifiError::InvalidData);
-        }
-        let mut config = Self::get_config(&env);
-        config.resolution_delay = delay;
-        env.storage().instance().set(&DataKey::Config, &config);
-        Self::extend_instance(&env);
-
-        ResolutionDelayUpdateEvent { admin, delay }.publish(&env);
-        Ok(())
-    }
-
-    /// Set claim window in seconds. Caller must have Admin role (0).
-    ///
-    /// The claim window defines how long after pool resolution users can claim winnings.
-    /// Must be between MIN_CLAIM_WINDOW (1 day) and MAX_CLAIM_WINDOW (365 days).
-    ///
-    /// # Arguments
-    /// * `admin` - Address with Admin role (0).
-    /// * `claim_window_seconds` - Claim window duration in seconds.
-    ///
-    /// # Errors
-    /// * `PredifiError::InvalidData` if claim_window_seconds is outside allowed range
-    /// * `PredifiError::Unauthorized` if caller doesn't have Admin role
-    pub fn set_claim_window(
-        env: Env,
-        admin: Address,
-        claim_window_seconds: u64,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "set_claim_window")?;
-
-        if !(MIN_CLAIM_WINDOW..=MAX_CLAIM_WINDOW).contains(&claim_window_seconds) {
-            return Err(PredifiError::InvalidData);
-        }
-
-        let mut config = Self::get_config(&env);
-        config.claim_window_seconds = claim_window_seconds;
-        env.storage().instance().set(&DataKey::Config, &config);
-        Self::extend_instance(&env);
-
-        ClaimWindowUpdateEvent {
-            admin,
-            claim_window_seconds,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Set minimum pool duration in seconds. Caller must have Admin role (0).
-    pub fn set_min_pool_duration(
-        env: Env,
-        admin: Address,
-        duration: u64,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "set_min_pool_duration")?;
-
-        let mut config = Self::get_config(&env);
-        config.min_pool_duration = duration;
-        env.storage().instance().set(&DataKey::Config, &config);
-        Self::extend_instance(&env);
-
-        MinPoolDurationUpdateEvent { admin, duration }.publish(&env);
-        Ok(())
-    }
-
-    /// Set the global minimum stake amount. Caller must have Admin role (0).
-    ///
-    /// Predictions with an amount below this threshold will be rejected with
-    /// `PredifiError::InsufficientStake`. This prevents spam from micro-predictions.
-    ///
-    /// # Arguments
-    /// * `admin`  - Address with Admin role (0).
-    /// * `amount` - New minimum stake in base token units. Must be > 0.
-    pub fn set_min_stake(env: Env, admin: Address, amount: i128) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "set_min_stake")?;
-        assert!(amount > 0, "min_stake must be greater than zero");
-
-        let mut config = Self::get_config(&env);
-        config.min_stake = amount;
-        env.storage().instance().set(&DataKey::Config, &config);
-        Self::extend_instance(&env);
-
-        MinStakeUpdateEvent {
-            admin,
-            min_stake: amount,
-        }
-        .publish(&env);
-        Ok(())
-    }
-
-    /// Set referral cut in basis points (e.g. 5000 = 50% of referrer's fee share). Caller must have Admin role (0).
-    /// Must be ≤ 10_000.
-    pub fn set_referral_cut_bps(
-        env: Env,
-        admin: Address,
-        referral_cut_bps: u32,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "set_referral_cut_bps")?;
-        assert!(
-            referral_cut_bps <= 10_000,
-            "referral_cut_bps must be at most 10000"
-        );
-        let mut config = Self::get_config(&env);
-        config.referral_bps = referral_cut_bps;
-        env.storage().instance().set(&DataKey::Config, &config);
-        env.storage()
-            .instance()
-            .set(&DataKey::ReferralCutBps, &referral_cut_bps);
-        Self::extend_instance(&env);
-        Ok(())
-    }
-
-    /// Set the referral reward rate in basis points stored in the Config struct.
-    ///
-    /// This allows admins to run "referral seasons" (e.g. raise from 500 bps / 5%
-    /// to 1000 bps / 10%) without any code changes.  The value is persisted in the
-    /// `Config` instance-storage entry so it is picked up automatically by fee
-    /// calculation logic.
-    ///
-    /// Caller must hold the Admin role. `bps` must be ≤ 10_000.
-    pub fn set_referral_rate(env: Env, admin: Address, bps: u32) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "set_referral_rate")?;
-        if bps > 10_000 {
-            return Err(PredifiError::InvalidFeeBps);
-        }
-        let mut config = Self::get_config(&env);
-        config.referral_bps = bps;
-        env.storage().instance().set(&DataKey::Config, &config);
-        Self::extend_instance(&env);
-        Ok(())
-    }
-
-    /// Add a token to the allowed betting whitelist. Caller must have Admin role (0).
-    pub fn add_token_to_whitelist(
-        env: Env,
-        admin: Address,
-        token: Address,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "add_token_to_whitelist")?;
-        let key = DataKey::TokenWl(token.clone());
-        env.storage().persistent().set(&key, &true);
-        Self::extend_persistent(&env, &key);
-
-        // Add to the whitelist list if not already present
-        let whitelist_key = DataKey::TokenWhitelist;
-        let mut whitelist: Vec<Address> = env
-            .storage()
-            .persistent()
-            .get(&whitelist_key)
-            .unwrap_or_else(|| Vec::new(&env));
-
-        // Only add if not already in the list
-        if !whitelist.contains(&token) {
-            whitelist.push_back(token.clone());
-            env.storage().persistent().set(&whitelist_key, &whitelist);
-            Self::extend_persistent(&env, &whitelist_key);
-        }
-
-        TokenWhitelistAddedEvent {
-            admin: admin.clone(),
-            token: token.clone(),
-        }
-        .publish(&env);
-        Ok(())
-    }
-
-    /// Remove a token from the allowed betting whitelist. Caller must have Admin role (0).
-    pub fn remove_token_from_whitelist(
-        env: Env,
-        admin: Address,
-        token: Address,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "remove_token_from_whitelist")?;
-        let key = DataKey::TokenWl(token.clone());
-        env.storage().persistent().remove(&key);
-
-        // Remove from the whitelist list
-        let whitelist_key = DataKey::TokenWhitelist;
-        let mut whitelist: Vec<Address> = env
-            .storage()
-            .persistent()
-            .get(&whitelist_key)
-            .unwrap_or_else(|| Vec::new(&env));
-
-        // Remove the token from the list if present
-        let new_whitelist = Vec::new(&env);
-        let mut new_whitelist = new_whitelist;
-        for t in whitelist.iter() {
-            if t.clone() != token {
-                new_whitelist.push_back(t);
-            }
-        }
-        whitelist = new_whitelist;
-
-        env.storage().persistent().set(&whitelist_key, &whitelist);
-        Self::extend_persistent(&env, &whitelist_key);
-
-        TokenWhitelistRemovedEvent {
-            admin: admin.clone(),
-            token: token.clone(),
-        }
-        .publish(&env);
-        Ok(())
-    }
-
-    /// Batch add multiple tokens to the allowed betting whitelist in a single
-    /// transaction. Caller must have Admin role (0).
-    ///
-    /// Reduces per-token transaction overhead when onboarding several tokens
-    /// at once. Skips tokens that are already whitelisted.
-    ///
-    /// # Errors
-    /// * `Unauthorized` - If caller does not hold the Admin role.
-    /// * `InvalidData` - If `tokens` is empty or exceeds `MAX_BATCH_SIZE` (100).
-    pub fn batch_add_tokens_to_whitelist(
-        env: Env,
-        admin: Address,
-        tokens: Vec<Address>,
-    ) -> Result<u32, PredifiError> {
-        const MAX_BATCH_SIZE: u32 = 100;
-
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "batch_add_tokens_to_whitelist")?;
-
-        let batch_size = tokens.len();
-        if batch_size == 0 || batch_size > MAX_BATCH_SIZE {
-            return Err(PredifiError::InvalidData);
-        }
-
-        let whitelist_key = DataKey::TokenWhitelist;
-        let mut whitelist: Vec<Address> = env
-            .storage()
-            .persistent()
-            .get(&whitelist_key)
-            .unwrap_or_else(|| Vec::new(&env));
-
-        let mut added_count: u32 = 0;
-        for token in tokens.iter() {
-            let key = DataKey::TokenWl(token.clone());
-            let already_whitelisted = env.storage().persistent().has(&key);
-
-            if !already_whitelisted {
-                env.storage().persistent().set(&key, &true);
-                Self::extend_persistent(&env, &key);
-
-                if !whitelist.contains(&token) {
-                    whitelist.push_back(token.clone());
-                }
-
-                added_count += 1;
-
-                TokenWhitelistAddedEvent {
-                    admin: admin.clone(),
-                    token: token.clone(),
-                }
-                .publish(&env);
-            }
-        }
-
-        if added_count > 0 {
-            env.storage().persistent().set(&whitelist_key, &whitelist);
-            Self::extend_persistent(&env, &whitelist_key);
-        }
-
-        Ok(added_count)
-    }
-
-    /// Batch remove multiple tokens from the allowed betting whitelist in a
-    /// single transaction. Caller must have Admin role (0).
-    ///
-    /// # Errors
-    /// * `Unauthorized` - If caller does not hold the Admin role.
-    /// * `InvalidData` - If `tokens` is empty or exceeds `MAX_BATCH_SIZE` (100).
-    pub fn batch_remove_tokens_from_whitelist(
-        env: Env,
-        admin: Address,
-        tokens: Vec<Address>,
-    ) -> Result<u32, PredifiError> {
-        const MAX_BATCH_SIZE: u32 = 100;
-
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "batch_remove_tokens_from_whitelist")?;
-
-        let batch_size = tokens.len();
-        if batch_size == 0 || batch_size > MAX_BATCH_SIZE {
-            return Err(PredifiError::InvalidData);
-        }
-
-        let whitelist_key = DataKey::TokenWhitelist;
-        let mut whitelist: Vec<Address> = env
-            .storage()
-            .persistent()
-            .get(&whitelist_key)
-            .unwrap_or_else(|| Vec::new(&env));
-
-        let mut removed_count: u32 = 0;
-        for token in tokens.iter() {
-            let key = DataKey::TokenWl(token.clone());
-            let was_whitelisted = env.storage().persistent().has(&key);
-
-            if was_whitelisted {
-                env.storage().persistent().remove(&key);
-                removed_count += 1;
-
-                TokenWhitelistRemovedEvent {
-                    admin: admin.clone(),
-                    token: token.clone(),
-                }
-                .publish(&env);
-            }
-        }
-
-        if removed_count > 0 {
-            let mut new_whitelist = Vec::new(&env);
-            for t in whitelist.iter() {
-                if !tokens.contains(&t) {
-                    new_whitelist.push_back(t);
-                }
-            }
-            whitelist = new_whitelist;
-
-            env.storage().persistent().set(&whitelist_key, &whitelist);
-            Self::extend_persistent(&env, &whitelist_key);
-        }
-
-        Ok(removed_count)
-    }
-
-    /// Add an oracle address to the trusted oracle whitelist. Caller must have Admin role (0).
-    pub fn add_oracle(
-        env: Env,
-        admin: Address,
-        oracle_address: Address,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "add_oracle")?;
-
-        let key = DataKey::OracleWl(oracle_address.clone());
-        env.storage().persistent().set(&key, &true);
-        Self::extend_persistent(&env, &key);
-
-        let list_key = DataKey::OracleWhitelist;
-        let mut whitelist: Vec<Address> = env
-            .storage()
-            .persistent()
-            .get(&list_key)
-            .unwrap_or_else(|| Vec::new(&env));
-
-        if !whitelist.contains(&oracle_address) {
-            whitelist.push_back(oracle_address.clone());
-            env.storage().persistent().set(&list_key, &whitelist);
-            Self::extend_persistent(&env, &list_key);
-        }
-
-        OracleWhitelistAddedEvent {
-            admin,
-            oracle: oracle_address,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Remove an oracle address from the trusted oracle whitelist. Caller must have Admin role (0).
-    pub fn remove_oracle(
-        env: Env,
-        admin: Address,
-        oracle_address: Address,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "remove_oracle")?;
-
-        let key = DataKey::OracleWl(oracle_address.clone());
-        env.storage().persistent().remove(&key);
-
-        let list_key = DataKey::OracleWhitelist;
-        let whitelist: Vec<Address> = env
-            .storage()
-            .persistent()
-            .get(&list_key)
-            .unwrap_or_else(|| Vec::new(&env));
-
-        let mut new_whitelist = Vec::new(&env);
-        for oracle in whitelist.iter() {
-            if oracle.clone() != oracle_address {
-                new_whitelist.push_back(oracle);
-            }
-        }
-
-        env.storage().persistent().set(&list_key, &new_whitelist);
-        Self::extend_persistent(&env, &list_key);
-
-        OracleWhitelistRemovedEvent {
-            admin,
-            oracle: oracle_address,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Get the list of all supported (whitelisted) tokens.
-    /// Returns a Vec of token addresses that are allowed for betting.
-    pub fn get_supported_tokens(env: Env) -> Vec<Address> {
-        let whitelist_key = DataKey::TokenWhitelist;
-        let whitelist: Vec<Address> = env
-            .storage()
-            .persistent()
-            .get(&whitelist_key)
-            .unwrap_or_else(|| Vec::new(&env));
-
-        if env.storage().persistent().has(&whitelist_key) {
-            Self::extend_persistent(&env, &whitelist_key);
-        }
-
-        whitelist
-    }
-
-    /// Upgrade the contract Wasm code. Only callable by Admin (role 0).
-    pub fn upgrade_contract(
-        env: Env,
-        admin: Address,
-        new_wasm_hash: BytesN<32>,
-    ) -> Result<(), PredifiError> {
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "upgrade_contract")?;
-
-        let old_version: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::Version)
-            .unwrap_or(0u32);
-        let new_version = old_version + 1;
-
-        env.deployer()
-            .update_current_contract_wasm(new_wasm_hash.clone());
-
-        env.storage()
-            .instance()
-            .set(&DataKey::Version, &new_version);
-        Self::extend_instance(&env);
-
-        UpgradeEvent {
-            admin: admin.clone(),
-            new_wasm_hash,
-        }
-        .publish(&env);
-
-        ContractUpgradedEvent {
-            old_version,
-            new_version,
-            upgraded_by: admin,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Post-upgrade migration logic.
-    ///
-    /// v2 migration: the deprecated `resolved` and `canceled` boolean fields have been
-    /// removed from the `Pool` struct. All state is now represented exclusively by the
-    /// `state: MarketState` field. Existing pools stored with the old schema are
-    /// automatically handled by Soroban's XDR codec — the removed fields are simply
-    /// ignored on read, so no explicit data rewrite is required.
-    pub fn migrate_state(env: Env, admin: Address) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "migrate_state")?;
-
-        // v2 migration: Add any state migration logic here.
-        // Use Self::validate_pool_invariants(&pool) to ensure pool data consistency
-        // during migrations.
-
-        Ok(())
-    }
-
-    /// Returns true if the given token is on the allowed betting whitelist.
-    pub fn is_token_allowed(env: Env, token: Address) -> bool {
-        Self::is_token_whitelisted(&env, &token)
-    }
-
-    /// Get referral cut in basis points (e.g. 5000 = 50% of referrer's fee share).
-    pub fn get_referral_cut_bps(env: Env) -> u32 {
-        Self::read_referral_cut_bps(&env)
-    }
-
-    /// Returns the current treasury and referral fee percentages as a [`FeeInfo`].
-    ///
-    /// - `treasury_fee_bps`: protocol fee charged on winnings (set via `set_fee_bps`).
-    /// - `referral_fee_bps`: share of the protocol fee paid to referrers (set via `set_referral_cut_bps`).
-    pub fn get_fees(env: Env) -> FeeInfo {
-        FeeInfo {
-            treasury_fee_bps: Self::get_config(&env).fee_bps,
-            referral_fee_bps: Self::read_referral_cut_bps(&env),
-        }
-    }
-
-    /// Returns the current cooldown in seconds between consecutive predictions from the same address.
-    pub fn get_prediction_cooldown(env: Env) -> u64 {
-        Self::get_config(&env).prediction_cooldown_seconds
-    }
-
     /// Returns true if the pool has a properly resolved outcome (not the sentinel value).
     fn is_pool_resolved(pool: &Pool) -> bool {
         pool.outcome != UNRESOLVED_OUTCOME
-    }
-
-    /// Return an aggregated metadata view of contract config and protocol state.
-    pub fn get_contract_info(env: Env) -> ContractInfo {
-        let config = Self::get_config(&env);
-        // If the access-control contract cannot be reached, fall back to the
-        // contract's own address so callers still receive a valid response.
-        let current_admin = Self::get_access_control_admin(&env, &config.access_control)
-            .unwrap_or_else(|_| env.current_contract_address());
-
-        ContractInfo {
-            version: env
-                .storage()
-                .instance()
-                .get(&DataKey::Version)
-                .unwrap_or(0u32),
-            current_admin,
-            is_paused: Self::is_paused(&env),
-            total_pools: env
-                .storage()
-                .instance()
-                .get(&DataKey::PoolIdCtr)
-                .unwrap_or(0u64),
-            fee_bps: config.fee_bps,
-            referral_cut_bps: Self::read_referral_cut_bps(&env),
-            treasury: config.treasury,
-            access_control: config.access_control,
-            resolution_delay: config.resolution_delay,
-            min_pool_duration: config.min_pool_duration,
-            min_stake: config.min_stake,
-            max_predictions_per_user: config.max_predictions_per_user,
-            prediction_cooldown_seconds: config.prediction_cooldown_seconds,
-        }
-    }
-
-    /// Get total referred volume for a (referrer, pool_id) in base token units.
-    pub fn get_referred_volume(env: Env, referrer: Address, pool_id: u64) -> i128 {
-        let key = DataKey::ReferredVolume(referrer, pool_id);
-        let vol = env.storage().persistent().get(&key).unwrap_or(0);
-        if env.storage().persistent().has(&key) {
-            Self::extend_persistent(&env, &key);
-        }
-        vol
-    }
-
-    /// Update or remove the referrer for a (user, pool_id) pair.
-    ///
-    /// Callable only by the user themselves. Allows correcting a mistaken or
-    /// compromised referrer address before or after predictions are placed.
-    ///
-    /// # Arguments
-    /// * `user`         - The user whose referrer is being updated (must provide auth).
-    /// * `pool_id`      - The pool for which the referrer is being updated.
-    /// * `new_referrer` - `Some(address)` to set a new referrer, `None` to remove it.
-    ///
-    /// # Errors
-    /// * `Unauthorized` if the caller is not the user.
-    pub fn update_referrer(
-        env: Env,
-        user: Address,
-        pool_id: u64,
-        new_referrer: Option<Address>,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        user.require_auth();
-        let referrer_key = DataKey::Referrer(user.clone(), pool_id);
-        match new_referrer {
-            Some(ref addr) => {
-                if addr == &user {
-                    return Err(PredifiError::Unauthorized);
-                }
-                env.storage().persistent().set(&referrer_key, addr);
-                Self::extend_persistent(&env, &referrer_key);
-            }
-            None => {
-                env.storage().persistent().remove(&referrer_key);
-            }
-        }
-
-        // Issue #1142: emit event so off-chain indexers stay in sync.
-        ReferrerUpdatedEvent {
-            user,
-            pool_id,
-            new_referrer,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Withdraw accumulated protocol fees or unused liquidity from the contract.
-    /// Only callable by Admin (role 0).
-    ///
-    /// # Arguments
-    /// * `admin` - Address with Admin role (must provide auth)
-    /// * `token` - The token contract address to withdraw
-    /// * `amount` - Amount to withdraw (must be > 0)
-    /// * `recipient` - Address to receive the withdrawn funds (typically treasury)
-    ///
-    /// # Returns
-    /// Result indicating success or error
-    ///
-    /// # Security
-    /// - Requires Admin role (0)
-    /// - Emits TreasuryWithdrawnEvent for audit trail
-    /// - Validates amount >= MIN_WITHDRAWAL_AMOUNT
-    /// - Checks contract has sufficient balance
-    pub fn withdraw_treasury(
-        env: Env,
-        admin: Address,
-        token: Address,
-        amount: i128,
-        recipient: Address,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-
-        // Verify admin role
-        Self::require_admin_role(&env, &admin, "withdraw_treasury")?;
-
-        // Reject zero or negative withdrawals before touching token state.
-        if amount <= 0 || amount < MIN_WITHDRAWAL_AMOUNT {
-            return Err(PredifiError::InvalidAmount);
-        }
-
-        // Get token client and check the contract's available balance first.
-        let token_client = token::Client::new(&env, &token);
-        let available_balance = token_client.balance(&env.current_contract_address());
-
-        // Verify sufficient balance
-        if available_balance < amount {
-            return Err(PredifiError::InsufficientBalance);
-        }
-
-        Self::enter_reentrancy_guard(&env);
-
-        // Validate token transfer before withdrawal
-        Self::validate_token_transfer(
-            &env,
-            &token,
-            &env.current_contract_address(),
-            &recipient,
-            amount,
-        )?;
-
-        // Transfer tokens to recipient
-        token_client.transfer(&env.current_contract_address(), &recipient, &amount);
-
-        // Compute remaining balance after transfer for the audit event
-        let remaining_balance = token_client.balance(&env.current_contract_address());
-
-        Self::exit_reentrancy_guard(&env);
-
-        // Emit audit event
-        TreasuryWithdrawnEvent {
-            admin: admin.clone(),
-            token: token.clone(),
-            amount,
-            recipient: recipient.clone(),
-            remaining_balance,
-            timestamp: env.ledger().timestamp(),
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Create a new prediction pool with configurable parameters.
-    ///
-    /// This function creates a new prediction market pool where users can stake tokens on
-    /// specific outcomes. The pool is initialized in the `Active` state and immediately
-    /// accepts predictions once the `start_time` is reached (or immediately if `start_time` is 0).
-    ///
-    /// # Parameters
-    ///
-    /// - `env` - The Soroban environment, providing access to storage, ledger, and auth
-    /// - `creator` - The address creating the pool (must authenticate via `require_auth`)
-    /// - `end_time` - Unix timestamp after which no more predictions are accepted (must be > current_time)
-    /// - `token` - The Stellar token contract address used for staking (must be whitelisted)
-    /// - `options_count` - Number of possible outcomes (must be >= 2 and <= MAX_OPTIONS_COUNT = 100)
-    /// - `category` - Market category symbol (e.g., `CATEGORY_SPORTS`, `CATEGORY_FINANCE`)
-    /// - `config` - Pool configuration parameters via `PoolConfig` struct:
-    ///   - `start_time` - Unix timestamp when the pool opens for predictions (0 = open immediately)
-    ///   - `description` - Short human-readable description (max 256 bytes)
-    ///   - `metadata_url` - URL to extended metadata, e.g., IPFS link (max 512 bytes)
-    ///   - `min_stake` - Minimum stake amount per prediction (must be > 0)
-    ///   - `max_stake` - Maximum stake per prediction (0 = no limit, else must be >= min_stake)
-    ///   - `min_total_stake` - Minimum total stake required for resolution (must be > 0)
-    ///   - `max_total_stake` - Hard cap on total stake (0 = no limit)
-    ///   - `initial_liquidity` - Optional house money provided by creator (must be >= 0)
-    ///   - `required_resolutions` - Number of oracle/operator votes needed for resolution (must be >= 1)
-    ///   - `private` - If true, only whitelisted addresses can participate
-    ///   - `whitelist_key` - Optional symbol for private pool access
-    ///   - `outcome_descriptions` - Human-readable labels for each outcome (length must equal options_count)
-    ///
-    /// # Return Value
-    ///
-    /// Returns `Ok(pool_id)` - The unique identifier of the newly created pool.
-    /// The pool_id is auto-incremented and can be used to reference the pool in all subsequent operations.
-    ///
-    /// # Emitted Events
-    ///
-    /// This function emits the following events:
-    ///
-    /// 1. **`PoolCreatedEvent`** (always emitted):
-    ///   - `pool_id` - The new pool's unique identifier
-    ///   - `creator` - The address that created the pool
-    ///   - `end_time` - The pool's betting deadline
-    ///   - `token` - The token contract used for staking
-    ///   - `options_count` - Number of possible outcomes
-    ///   - `metadata_url` - URL to extended metadata
-    ///   - `initial_liquidity` - Amount of house money provided
-    ///   - `category` - Market category
-    ///   - `required_resolutions` - Resolution threshold
-    ///   - `max_total_stake` - Total stake cap
-    ///   - `outcome_descriptions` - Labels for each outcome
-    ///
-    /// 2. **`InitialLiquidityProvidedEvent`** (conditional):
-    ///   - Emitted only if `initial_liquidity > 0`
-    ///   - `pool_id` - The pool receiving liquidity
-    ///   - `creator` - The address providing liquidity
-    ///   - `amount` - The amount of liquidity transferred
-    ///
-    /// # Error Conditions
-    ///
-    /// The function can return the following errors:
-    ///
-    /// - `ContractPaused` - The contract is currently paused; all state-mutating operations are blocked
-    /// - `Unauthorized` - Caller is not authorized (creator authentication failed)
-    /// - `TokenNotWhitelisted` - The specified token is not on the allowed betting whitelist
-    /// - `InvalidTimestamp` - `end_time` is not in the future, exceeds MAX_POOL_DURATION, or `end_time <= start_time`
-    /// - `DeadlineInPast` - `end_time` or `start_time` is in the past (issue #1130)
-    /// - `InvalidData` - `options_count` < 2 or > MAX_OPTIONS_COUNT
-    /// - `MetadataUrlInvalid` - `metadata_url` exceeds 512 bytes
-    /// - `InvalidTargetPrice` - Invalid target price (for price-based pools)
-    /// - `InitialLiquidityBelowSafetyMargin` - Initial liquidity is insufficient relative to `max_total_stake` (issue #1131)
-    /// - `RequiredResolutionsExceedOperators` - `required_resolutions` > number of active operators
-    /// - `OutcomeDescriptionTooLong` - An outcome description exceeds MAX_OUTCOME_DESCRIPTION_LEN (128 bytes)
-    /// - `OutcomeDescriptionEmpty` - An outcome description is empty or below MIN_OUTCOME_DESCRIPTION_LEN (1 byte)
-    ///
-    /// # Validation Rules
-    ///
-    /// The function performs extensive validation before pool creation:
-    ///
-    /// **Time Validation:**
-    /// - `end_time` must be > current_time (INV-8)
-    /// - `end_time` must be > `start_time`
-    /// - `end_time` must not exceed current_time + MAX_POOL_DURATION (365 days)
-    /// - Pool duration must be >= min_pool_duration (default: 1 hour)
-    /// - If `start_time > 0`, it must be > current_time (cannot schedule past start)
-    ///
-    /// **Token Validation:**
-    /// - Token must be whitelisted via `is_token_whitelisted`
-    /// - Token address validation is deferred to actual transfer (Soroban pattern)
-    ///
-    /// **Outcome Validation:**
-    /// - `options_count` must be >= 2 (binary or multi-outcome)
-    /// - `options_count` must be <= MAX_OPTIONS_COUNT (100)
-    /// - `outcome_descriptions` length must equal `options_count`
-    /// - Each outcome description must be between 1 and 128 bytes
-    ///
-    /// **Stake Validation:**
-    /// - `min_stake` must be > 0
-    /// - `max_stake` must be 0 or >= `min_stake`
-    /// - `min_total_stake` must be > 0
-    /// - `max_total_stake` must be >= 0
-    ///
-    /// **Liquidity Validation:**
-    /// - `initial_liquidity` must be >= 0
-    /// - `initial_liquidity` must not exceed MAX_INITIAL_LIQUIDITY
-    /// - If `max_total_stake > 0` and `initial_liquidity > 0`: liquidity must be at least
-    ///   `(max_total_stake * INITIAL_LIQUIDITY_SAFETY_MARGIN_BPS) / 10_000` (1% safety margin)
-    ///
-    /// **Resolution Validation:**
-    /// - `required_resolutions` must be >= 1
-    /// - `required_resolutions` must not exceed the number of active operators
-    ///   (unless operator_count is 0, in which case oracle resolution is allowed)
-    ///
-    /// **Category Validation:**
-    /// - Category must be one of the canonical category symbols (e.g., CATEGORY_SPORTS)
-    ///
-    /// **Private Pool Validation:**
-    /// - If `whitelist_key` is provided, it must pass `validate_referral_code`
-    ///
-    /// # Storage Initialization
-    ///
-    /// The function initializes the following storage entries:
-    ///
-    /// - `DataKey::Pool(pool_id)` - The pool's full state
-    /// - `DataKey::PartCnt(pool_id)` - Participant count (initialized to 0)
-    /// - `DataKey::OutStakes(pool_id)` - Batch storage for outcome stakes (initialized to zeros)
-    /// - `DataKey::CatPoolCt(category)` - Category pool count (incremented)
-    /// - `DataKey::CatPoolIx(category, index)` - Category pool index (appended)
-    /// - `DataKey::PoolIdCtr` - Pool ID counter (incremented)
-    ///
-    /// # Initial Liquidity Transfer
-    ///
-    /// If `initial_liquidity > 0`, the function transfers tokens from the creator to the contract:
-    /// - Tokens are transferred via `token_client.transfer(creator, contract, initial_liquidity)`
-    /// - This represents "house money" that participates in the pool
-    /// - Initial liquidity is part of `pool.total_stake` but typically excluded from fee calculations
-    /// - The transfer occurs after all validation to avoid wasting gas on invalid pools
-    ///
-    /// # Category Indexing
-    ///
-    /// Pools are indexed by category for efficient querying:
-    /// - Each category maintains a count of pools (`CatPoolCt`)
-    /// - Each category maintains an ordered list of pool IDs (`CatPoolIx`)
-    /// - This enables category-based pool enumeration and discovery
-    ///
-    /// # Pre-conditions
-    ///
-    /// - `end_time > current_time` (INV-8)
-    /// - Token must be whitelisted
-    /// - Creator must have sufficient balance for `initial_liquidity` (if provided)
-    /// - Access control contract must be deployed and have sufficient operators (if required_resolutions > 0)
-    ///
-    /// # Post-conditions
-    ///
-    /// - `Pool.state = Active`
-    /// - `Pool.total_stake = initial_liquidity` (if provided, else 0)
-    /// - `Pool.outcome = UNRESOLVED_OUTCOME`
-    /// - `PoolIdCtr` is incremented
-    /// - Category indexes are updated
-    /// - `PoolCreatedEvent` is emitted
-    /// - `InitialLiquidityProvidedEvent` is emitted (if liquidity provided)
-    ///
-    /// # Usage Examples
-    ///
-    /// ## Basic Binary Pool
-    ///
-    /// ```rust
-    /// use soroban_sdk::{Address, Symbol, symbol_short};
-    ///
-    /// let creator = Address::generate(&env);
-    /// let token = Address::generate(&env);
-    /// let end_time = env.ledger().timestamp() + 86400; // 24 hours from now
-    ///
-    /// let config = PoolConfig {
-    ///     start_time: 0, // Open immediately
-    ///     description: String::from_str(&env, "Will BTC exceed $100k?"),
-    ///     metadata_url: String::from_str(&env, "ipfs://QmHash"),
-    ///     min_stake: 1000,
-    ///     max_stake: 1000000,
-    ///     min_total_stake: 10000,
-    ///     max_total_stake: 10000000,
-    ///     initial_liquidity: 0,
-    ///     required_resolutions: 1,
-    ///     private: false,
-    ///     whitelist_key: None,
-    ///     outcome_descriptions: vec![&env, 
-    ///         String::from_str(&env, "No"),
-    ///         String::from_str(&env, "Yes")
-    ///     ],
-    /// };
-    ///
-    /// let pool_id = contract.create_pool(
-    ///     env,
-    ///     creator,
-    ///     end_time,
-    ///     token,
-    ///     2, // Binary: No/Yes
-    ///     CATEGORY_CRYPTO,
-    ///     config,
-    /// )?;
-    /// ```
-    ///
-    /// ## Pool with Initial Liquidity
-    ///
-    /// ```rust
-    /// let config = PoolConfig {
-    ///     // ... other fields ...
-    ///     initial_liquidity: 1000000, // Creator provides house money
-    ///     max_total_stake: 10000000,
-    ///     // ... other fields ...
-    /// };
-    ///
-    /// let pool_id = contract.create_pool(env, creator, end_time, token, 2, CATEGORY_SPORTS, config)?;
-    /// // Creator must have approved the contract to spend 1000000 tokens
-    /// ```
-    ///
-    /// ## Private Pool with Whitelist
-    ///
-    /// ```rust
-    /// let config = PoolConfig {
-    ///     // ... other fields ...
-    ///     private: true,
-    ///     whitelist_key: Some(symbol_short!("SECRET_KEY")),
-    ///     // ... other fields ...
-    /// };
-    ///
-    /// let pool_id = contract.create_pool(env, creator, end_time, token, 2, CATEGORY_FINANCE, config)?;
-    /// // Only whitelisted users or those with the invite key can participate
-    /// ```
-    #[allow(clippy::too_many_arguments)]
-    pub fn create_pool(
-        env: Env,
-        creator: Address,
-        end_time: u64,
-        token: Address,
-        options_count: u32,
-        category: Symbol,
-        config: PoolConfig,
-    ) -> Result<u64, PredifiError> {
-        Self::require_not_paused(&env)?;
-        creator.require_auth();
-
-        // Validate: category must be in the allowed list, return error if invalid
-        let normalized_category = match Self::validate_category(&env, &category) {
-            Ok(cat) => cat,
-            Err(e) => soroban_sdk::panic_with_error!(&env, e),
-        };
-
-        // Validate: token must be on the allowed betting whitelist
-        if !Self::is_token_whitelisted(&env, &token) {
-            soroban_sdk::panic_with_error!(&env, PredifiError::TokenNotWhitelisted);
-        }
-
-        let current_time = env.ledger().timestamp();
-
-        // Validate: end_time must be greater than start_time
-        if end_time <= config.start_time {
-            soroban_sdk::panic_with_error!(&env, PredifiError::InvalidTimestamp);
-        }
-
-        // Issue #1130 — staking deadlines must be in the future. End_time
-        // strictly in the past raises the `DeadlineInPast` protocol error
-        // (typed, machine-checkable). The legacy `assert!` below still
-        // catches the boundary case `end_time == current_time` with its
-        // diagnostic message for backwards-compatible test expectations.
-        // A `start_time` of 0 is a sentinel for "open immediately"; any
-        // non-zero start_time must be strictly in the future so a creator
-        // cannot schedule a pool that opens in the past.
-        if end_time < current_time {
-            soroban_sdk::panic_with_error!(&env, PredifiError::DeadlineInPast);
-        }
-        if config.start_time > 0 && config.start_time < current_time {
-            soroban_sdk::panic_with_error!(&env, PredifiError::DeadlineInPast);
-        }
-
-        // Validate: end_time must be in the future (legacy assert kept for
-        // diagnostic clarity; the DeadlineInPast check above is the
-        // authoritative protocol error).
-        assert!(end_time > current_time, "end_time must be in the future");
-
-        // Validate: end_time must not exceed MAX_POOL_DURATION from now
-        if end_time > current_time + MAX_POOL_DURATION {
-            soroban_sdk::panic_with_error!(&env, PredifiError::InvalidTimestamp);
-        }
-
-        let min_pool_duration = env
-            .storage()
-            .instance()
-            .get::<DataKey, Config>(&DataKey::Config)
-            .map(|c| c.min_pool_duration)
-            .unwrap_or(DEFAULT_MIN_POOL_DURATION);
-
-        // Validate: minimum pool duration
-        assert!(
-            end_time >= current_time + min_pool_duration,
-            "end_time must be at least min_pool_duration in the future"
-        );
-
-        // Validate: options_count must be at least 2 (binary or more outcomes)
-        if options_count < 2 {
-            return Err(PredifiError::InvalidData);
-        }
-
-        // Validate: options_count must not exceed maximum limit
-        if options_count > MAX_OPTIONS_COUNT {
-            return Err(PredifiError::InvalidData);
-        }
-
-        // Validate: initial_liquidity must be non-negative if provided
-        assert!(
-            config.initial_liquidity >= 0,
-            "initial_liquidity must be non-negative"
-        );
-
-        // Validate: initial_liquidity must not exceed maximum limit
-        assert!(
-            config.initial_liquidity <= MAX_INITIAL_LIQUIDITY,
-            "initial_liquidity exceeds maximum allowed value"
-        );
-
-        // Issue #1131 — initial-liquidity safety margin. When the creator
-        // caps the pool at `max_total_stake > 0`, the seeded liquidity must
-        // cover at least INITIAL_LIQUIDITY_SAFETY_MARGIN_BPS basis points of
-        // that cap, so that early high-value predictions cannot drain a
-        // thinly-seeded pool faster than it can be cancelled.
-        if config.max_total_stake > 0 && config.initial_liquidity > 0 {
-            // safety_min = max_total_stake * bps / 10_000 — saturating to
-            // avoid arithmetic panics on extreme inputs.
-            let bps = INITIAL_LIQUIDITY_SAFETY_MARGIN_BPS as i128;
-            let safety_min = config
-                .max_total_stake
-                .checked_mul(bps)
-                .map(|v| v / 10_000)
-                .unwrap_or(i128::MAX);
-            if config.initial_liquidity < safety_min {
-                soroban_sdk::panic_with_error!(
-                    &env,
-                    PredifiError::InitialLiquidityBelowSafetyMargin
-                );
-            }
-        }
-
-        // Validate: required_resolutions must be at least 1
-        assert!(
-            config.required_resolutions >= 1,
-            "required_resolutions must be at least 1"
-        );
-
-        // Validate: required_resolutions must not exceed the number of active operators.
-        // If required_resolutions > operator_count, the pool can never reach the resolution
-        // threshold and will be permanently stuck in the Active state.
-        // WARNING: This is a hard check — pool creation will fail if there are not enough
-        // operators registered in the access_control contract to satisfy required_resolutions.
-        // Note: If operator_count is 0, the pool can still be resolved by oracles.
-        {
-            let cfg = Self::get_config(&env);
-            // Use try_invoke_contract so that an unreachable access-control contract maps
-            // to OracleNotInitialized rather than causing an unhandled panic.
-            let operator_count: u32 = env
-                .try_invoke_contract::<u32, PredifiError>(
-                    &cfg.access_control,
-                    &Symbol::new(&env, "get_operator_count"),
-                    soroban_sdk::vec![&env],
-                )
-                .map_err(|_| PredifiError::OracleNotInitialized)
-                .and_then(|inner| inner.map_err(|_| PredifiError::OracleNotInitialized))
-                .unwrap_or(0u32); // default: treat as zero operators so creation can proceed
-            if operator_count > 0 && config.required_resolutions > operator_count {
-                soroban_sdk::panic_with_error!(
-                    &env,
-                    PredifiError::RequiredResolutionsExceedOperators
-                );
-            }
-        }
-
-        // Note: Token address validation is deferred to when the token is actually used.
-        // This is the standard pattern in Soroban - invalid tokens will fail when
-        // transfers are attempted during place_prediction.
-
-        assert!(
-            config.description.len() <= 256,
-            "description exceeds 256 bytes"
-        );
-        if config.metadata_url.len() > 512 {
-            soroban_sdk::panic_with_error!(&env, PredifiError::MetadataUrlInvalid);
-        }
-
-        // Validate stake limits
-        assert!(config.min_stake > 0, "min_stake must be greater than zero");
-        assert!(
-            config.max_stake == 0 || config.max_stake >= config.min_stake,
-            "max_stake must be zero (unlimited) or >= min_stake"
-        );
-        // Validate: min_total_stake must be strictly positive (> 0)
-        assert!(
-            config.min_total_stake > 0,
-            "min_total_stake must be greater than zero"
-        );
-        assert!(config.max_total_stake >= 0, "max_total_stake must be >= 0");
-
-        if let Some(ref whitelist_key) = config.whitelist_key {
-            if let Err(e) = Self::validate_referral_code(&env, whitelist_key) {
-                soroban_sdk::panic_with_error!(&env, e);
-            }
-        }
-
-        // outcome_descriptions validation is now handled by validate_pool_invariants
-        // called right after pool structure is initialized.
-
-        let pool_id: u64 = env
-            .storage()
-            .instance()
-            .get(&DataKey::PoolIdCtr)
-            .unwrap_or(0);
-        // Initialize pool data structure
-        let pool = Pool {
-            start_time: config.start_time,
-            end_time,
-            state: MarketState::Active,
-            outcome: UNRESOLVED_OUTCOME,
-            token: token.clone(),
-            total_stake: config.initial_liquidity,
-            category: normalized_category,
-            description: config.description.clone(),
-            metadata_url: config.metadata_url.clone(),
-            options_count,
-            min_stake: config.min_stake,
-            max_stake: config.max_stake,
-            min_total_stake: config.min_total_stake,
-            max_total_stake: config.max_total_stake,
-            initial_liquidity: config.initial_liquidity,
-            creator: creator.clone(),
-            required_resolutions: config.required_resolutions,
-            private: config.private,
-            whitelist_key: config.whitelist_key.clone(),
-            outcome_descriptions: config.outcome_descriptions.clone(),
-            fee_bps: 0, // Will be set at resolution
-            participants_count: 0,
-            resolution_timestamp: None, // Set when pool is resolved
-        };
-
-        Self::validate_pool_invariants(&pool);
-
-        let pool_key = DataKey::Pool(pool_id);
-        env.storage().persistent().set(&pool_key, &pool);
-        Self::bump_ttl(&env, &pool_key);
-
-        // Initialize optimized batch storage with zeros to avoid expensive fallback reads
-        let initial_stakes = gas_opt::alloc_zero_stakes(&env, options_count);
-        let stakes_key = DataKey::OutStakes(pool_id);
-        env.storage().persistent().set(&stakes_key, &initial_stakes);
-        Self::extend_persistent(&env, &stakes_key);
-
-        // Transfer initial liquidity from creator to contract if provided
-        if config.initial_liquidity > 0 {
-            let token_client = token::Client::new(&env, &token);
-            token_client.transfer(
-                &creator,
-                env.current_contract_address(),
-                &config.initial_liquidity,
-            );
-        }
-
-        // Update category index
-        let category_count_key = DataKey::CatPoolCt(category.clone());
-        let category_count: u32 = env
-            .storage()
-            .persistent()
-            .get(&category_count_key)
-            .unwrap_or(0);
-
-        let category_index_key = DataKey::CatPoolIx(category.clone(), category_count);
-        env.storage()
-            .persistent()
-            .set(&category_index_key, &pool_id);
-        Self::bump_ttl(&env, &category_index_key);
-
-        env.storage()
-            .persistent()
-            .set(&category_count_key, &(category_count + 1));
-        Self::bump_ttl(&env, &category_count_key);
-
-        env.storage()
-            .instance()
-            .set(&DataKey::PoolIdCtr, &(pool_id + 1));
-        Self::extend_instance(&env);
-
-        PoolCreatedEvent {
-            pool_id,
-            creator: creator.clone(),
-            end_time,
-            token,
-            options_count,
-            metadata_url: config.metadata_url,
-            initial_liquidity: config.initial_liquidity,
-            category,
-            required_resolutions: config.required_resolutions,
-            max_total_stake: config.max_total_stake,
-            outcome_descriptions: config.outcome_descriptions,
-        }
-        .publish(&env);
-
-        // Emit initial liquidity event if liquidity was provided
-        if config.initial_liquidity > 0 {
-            InitialLiquidityProvidedEvent {
-                pool_id,
-                creator,
-                amount: config.initial_liquidity,
-            }
-            .publish(&env);
-        }
-
-        // Register pool in the global active pool index.
-        Self::add_to_active_index(&env, pool_id);
-
-        Ok(pool_id)
-    }
-
-    /// Increase the maximum total stake cap for a pool.
-    /// Only the pool creator can increase it, and only before the market ends.
-    ///
-    /// - `new_max_total_stake` must be >= current `pool.total_stake`.
-    /// - Setting to 0 means "no cap" (only allowed if current cap is 0 or increasing from a non-zero).
-    pub fn increase_max_total_stake(
-        env: Env,
-        creator: Address,
-        pool_id: u64,
-        new_max_total_stake: i128,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        creator.require_auth();
-
-        let pool_key = DataKey::Pool(pool_id);
-        let mut pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-        Self::extend_persistent(&env, &pool_key);
-
-        if pool.creator != creator {
-            return Err(PredifiError::Unauthorized);
-        }
-
-        // Pool must still be active and not ended
-        // if pool.state != MarketState::Active {
-        //     return Err(PredifiError::InvalidPoolState);
-        // }
-        if !Self::is_pool_active(&pool) {
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        assert!(env.ledger().timestamp() < pool.end_time, "Pool has ended");
-
-        // Must not set a cap below what is already staked
-        assert!(
-            new_max_total_stake == 0 || new_max_total_stake >= pool.total_stake,
-            "new_max_total_stake must be zero (unlimited) or >= total_stake"
-        );
-
-        // Only allow increasing the cap (or setting unlimited)
-        if pool.max_total_stake > 0 && new_max_total_stake > 0 {
-            assert!(
-                new_max_total_stake >= pool.max_total_stake,
-                "new_max_total_stake must be >= current max_total_stake"
-            );
-        }
-
-        pool.max_total_stake = new_max_total_stake;
-        env.storage().persistent().set(&pool_key, &pool);
-        Self::extend_persistent(&env, &pool_key);
-
-        // Issue #1142: emit event so frontends/indexers can update the displayed cap.
-        MaxTotalStakeIncreasedEvent {
-            pool_id,
-            creator,
-            new_max_total_stake,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Update the description of a pool before any participant has joined.
-    ///
-    /// Allows the pool creator or a protocol admin to correct a typo or clarify
-    /// ambiguous wording. Once the first prediction is placed the description is
-    /// locked to prevent fraud.
-    ///
-    /// # Arguments
-    /// * `caller`   - Pool creator **or** an address with Admin role (0).
-    /// * `pool_id`  - The pool to update.
-    /// * `new_desc` - Replacement description (max 256 bytes, must be non-empty).
-    ///
-    /// # Errors
-    /// * `Unauthorized`     – caller is neither the creator nor an admin.
-    /// * `InvalidPoolState` – pool is not `Active`, has ended, or already has participants.
-    /// * `InvalidAmount`    – description is empty or exceeds 256 bytes.
-    pub fn update_pool_description(
-        env: Env,
-        caller: Address,
-        pool_id: u64,
-        new_desc: String,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        caller.require_auth();
-
-        let pool_key = DataKey::Pool(pool_id);
-        let mut pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-        Self::validate_pool_invariants(&pool);
-        Self::extend_persistent(&env, &pool_key);
-
-        // Only the creator or a protocol admin may update the description.
-        let is_creator = pool.creator == caller;
-        let is_admin = Self::require_admin_role(&env, &caller, "update_pool_description").is_ok();
-        if !is_creator && !is_admin {
-            return Err(PredifiError::Unauthorized);
-        }
-
-        // Pool must still be active (not resolved or canceled).
-        if !Self::is_pool_active(&pool) {
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        // Pool must not have ended.
-        if env.ledger().timestamp() >= pool.end_time {
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        // Lock the description once any participant has joined — equivalent to
-        // "pool has started" in this contract's model (no separate start_time).
-        // We read the pool's participants_count field which is the authoritative participant counter.
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env.storage().persistent().get(&pool_key).expect("Pool not found");
-        if pool.participants_count > 0 {
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        // Validate the new description: non-empty and within the 256-byte limit.
-        if new_desc.is_empty() || new_desc.len() > 256 {
-            return Err(PredifiError::InvalidAmount);
-        }
-
-        pool.description = new_desc.clone();
-        env.storage().persistent().set(&pool_key, &pool);
-        Self::extend_persistent(&env, &pool_key);
-
-        PoolDescriptionUpdatedEvent {
-            pool_id,
-            caller,
-            new_description: new_desc,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Finalises a prediction pool by recording an operator's vote for a winning outcome.
-    ///
-    /// Resolution uses a **multi-vote / threshold model**: each address with Operator
-    /// role (role `1`) calls this function once to cast a vote.  The pool transitions
-    /// to [`MarketState::Resolved`] only when a single outcome accumulates at least
-    /// `pool.required_resolutions` votes.  If operators disagree on the outcome a
-    /// [`PredifiError::ResolutionConflict`] is returned and the pool stays active so
-    /// administrators can intervene.
-    ///
-    /// # Resolution flow
-    ///
-    /// 1. **Auth & role check** — `operator` must sign the transaction and hold the
-    ///    Operator role for `pool_id` (`require_operator_role_for_resolution`).
-    /// 2. **State guard** — the pool must be in `MarketState::Active`; Locked,
-    ///    Resolved, and Cancelled pools are rejected with [`PredifiError::InvalidPoolState`].
-    /// 3. **Resolution delay** — `current_ledger_time >= pool.end_time + config.resolution_delay`
-    ///    must hold.  This cooling-off period allows late price feeds to settle before
-    ///    any outcome is locked in.  Violations return [`PredifiError::ResolutionDelayNotMet`].
-    /// 4. **Outcome validation** — `outcome` must satisfy
-    ///    `0 <= outcome < pool.options_count` and must not equal `UNRESOLVED_OUTCOME`
-    ///    (the sentinel value reserved for pools that have not yet been decided).
-    ///    Invalid values return [`PredifiError::InvalidOutcome`].
-    /// 5. **Duplicate-vote guard** — each operator may vote exactly once per pool.
-    ///    A second call from the same address returns [`PredifiError::OracleAlreadyVoted`].
-    /// 6. **Vote recording** — the vote is stored in temporary ledger storage under
-    ///    `DataKey::ResVote(pool_id, operator)`.  Per-outcome tallies are maintained
-    ///    in `DataKey::ResVoteCt(pool_id, outcome)` and the total vote count in
-    ///    `DataKey::ResTotal(pool_id)`.
-    /// 7. **Conflict detection** — if any other outcome already has votes, a
-    ///    [`ResolutionConflictEvent`] is emitted and [`PredifiError::ResolutionConflict`]
-    ///    is returned, leaving the pool in `Active` so administrators can adjudicate.
-    /// 8. **Threshold check & finalisation** — once `new_outcome_votes >= pool.required_resolutions`:
-    ///    - `pool.state` is set to `MarketState::Resolved`.
-    ///    - `pool.outcome` is recorded.
-    ///    - A dynamic protocol fee (`pool.fee_bps`) is calculated via
-    ///      [`Self::calculate_dynamic_fee`].
-    ///    - `pool.resolution_timestamp` is stamped with the current ledger time.
-    ///    - The pool is removed from the global active index.
-    ///    - [`PoolResolvedEvent`] and [`PoolResolvedDiagEvent`] are published.
-    ///
-    /// # Payout calculation (post-resolution)
-    ///
-    /// This function does **not** distribute funds directly.  Winners call
-    /// `claim_winnings` (or a payout helper) after resolution.  The payout for a
-    /// winning stake `s` out of a total winning-side stake `W` and a total pool
-    /// stake `T` is:
-    ///
-    /// ```text
-    /// gross_payout = s * T / W
-    /// fee          = gross_payout * pool.fee_bps / 10_000
-    /// net_payout   = gross_payout - fee
-    /// ```
-    ///
-    /// # Emitted events
-    ///
-    /// | Event | When |
-    /// |---|---|
-    /// | [`ResolutionVoteCastEvent`] | Every successful vote, regardless of threshold |
-    /// | [`ResolutionConflictEvent`] | When a vote conflicts with an earlier outcome |
-    /// | [`PoolResolvedEvent`] | When the threshold is reached and the pool is finalised |
-    /// | [`PoolResolvedDiagEvent`] | Same trigger as `PoolResolvedEvent`; carries stake diagnostics |
-    ///
-    /// # Errors
-    ///
-    /// | Error | Condition |
-    /// |---|---|
-    /// | [`PredifiError::ContractPaused`] | The contract is globally paused |
-    /// | [`PredifiError::InvalidPoolState`] | Pool is not `Active` |
-    /// | [`PredifiError::ResolutionDelayNotMet`] | Too early to resolve |
-    /// | [`PredifiError::InvalidOutcome`] | `outcome >= options_count` or equals sentinel |
-    /// | [`PredifiError::OracleAlreadyVoted`] | This operator already voted for this pool |
-    /// | [`PredifiError::ResolutionConflict`] | Operators disagree on the winning outcome |
-    ///
-    /// # Preconditions
-    ///
-    /// - `pool.state == MarketState::Active`
-    /// - `operator` holds Operator role (`1`) for `pool_id`
-    /// - `env.ledger().timestamp() >= pool.end_time + config.resolution_delay`
-    ///
-    /// # Postconditions (on threshold reached)
-    ///
-    /// - `pool.state == MarketState::Resolved` (INV-2 state transition satisfied)
-    /// - `pool.outcome == outcome`
-    /// - `pool.resolution_timestamp == Some(current_time)`
-    /// - Pool is removed from the global active index
-    pub fn resolve_pool(
-        env: Env,
-        operator: Address,
-        pool_id: u64,
-        outcome: u32,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        operator.require_auth();
-        Self::require_operator_role_for_resolution(&env, &operator, pool_id)?;
-
-        let pool_key = DataKey::Pool(pool_id);
-        let mut pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-
-        Self::validate_pool_invariants(&pool);
-
-        // if pool.state != MarketState::Active {
-        //     return Err(PredifiError::InvalidPoolState);
-        // }
-        if !Self::is_pool_active(&pool) {
-            log!(
-                &env,
-                "resolve_pool rejected: pool is not active",
-                pool_id,
-                operator.clone(),
-                outcome,
-                pool.end_time
-            );
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        let current_time = env.ledger().timestamp();
-        let config = Self::get_config(&env);
-        let eligible_at = pool.end_time.saturating_add(config.resolution_delay);
-
-        if current_time < eligible_at {
-            log!(
-                &env,
-                "resolve_pool rejected: resolution delay not met",
-                pool_id,
-                operator.clone(),
-                outcome,
-                current_time,
-                eligible_at
-            );
-            return Err(PredifiError::ResolutionDelayNotMet);
-        }
-
-        // Validate: outcome must be within the valid options range
-        if outcome >= pool.options_count {
-            log!(
-                &env,
-                "resolve_pool rejected: outcome is out of bounds",
-                pool_id,
-                operator.clone(),
-                outcome,
-                pool.options_count
-            );
-            return Err(PredifiError::InvalidOutcome);
-        }
-
-        // Validate: outcome cannot be the sentinel value
-        if outcome == UNRESOLVED_OUTCOME {
-            log!(
-                &env,
-                "resolve_pool rejected: outcome cannot be sentinel value",
-                pool_id,
-                operator.clone(),
-                outcome
-            );
-            return Err(PredifiError::InvalidOutcome);
-        }
-
-        // --- Multi-resolution Voting Logic ---
-
-        // Check if this operator has already voted for this pool
-        let vote_key = DataKey::ResVote(pool_id, operator.clone());
-        if env.storage().temporary().has(&vote_key) {
-            log!(
-                &env,
-                "resolve_pool rejected: operator already voted",
-                pool_id,
-                operator.clone(),
-                outcome
-            );
-            return Err(PredifiError::OracleAlreadyVoted); // Reusing error code for operators
-        }
-
-        // Record the operator's vote in temporary storage
-        env.storage().temporary().set(&vote_key, &outcome);
-        Self::extend_temporary(&env, &vote_key);
-
-        // Increment total number of votes cast for this pool
-        let total_votes_key = DataKey::ResTotal(pool_id);
-        let total_votes: u32 = env.storage().temporary().get(&total_votes_key).unwrap_or(0);
-        let new_total_votes = total_votes + 1;
-        env.storage()
-            .temporary()
-            .set(&total_votes_key, &new_total_votes);
-        Self::extend_temporary(&env, &total_votes_key);
-
-        // Increment specific outcome vote count
-        let outcome_votes_key = DataKey::ResVoteCt(pool_id, outcome);
-        let outcome_votes: u32 = env
-            .storage()
-            .temporary()
-            .get(&outcome_votes_key)
-            .unwrap_or(0);
-        let new_outcome_votes = outcome_votes + 1;
-        env.storage()
-            .temporary()
-            .set(&outcome_votes_key, &new_outcome_votes);
-        Self::extend_temporary(&env, &outcome_votes_key);
-
-        // Emit a ResolutionVoteCastEvent for observability
-        ResolutionVoteCastEvent {
-            pool_id,
-            voter: operator.clone(),
-            outcome,
-            vote_count: new_outcome_votes,
-            required_resolutions: pool.required_resolutions,
-        }
-        .publish(&env);
-
-        // Detect conflicts
-        if new_total_votes > new_outcome_votes {
-            for i in 0..pool.options_count {
-                if i == outcome {
-                    continue;
-                }
-                let other_key = DataKey::ResVoteCt(pool_id, i);
-                if env.storage().temporary().has(&other_key) {
-                    ResolutionConflictEvent {
-                        pool_id,
-                        oracle: operator.clone(),
-                        outcome,
-                        existing_outcome: i,
-                    }
-                    .publish(&env);
-                    return Err(PredifiError::ResolutionConflict);
-                }
-            }
-        }
-
-        // Check if the required threshold has been met
-        if new_outcome_votes >= pool.required_resolutions {
-            pool.state = MarketState::Resolved;
-            pool.outcome = outcome;
-            pool.fee_bps = Self::calculate_dynamic_fee(&env, &pool);
-            pool.resolution_timestamp = Some(env.ledger().timestamp()); // Record resolution time
-
-            env.storage().persistent().set(&pool_key, &pool);
-
-            // Remove from global active index now that the pool is resolved.
-            Self::remove_from_active_index(&env, pool_id);
-            Self::bump_ttl(&env, &pool_key);
-
-            // Retrieve winning-outcome stake for the diagnostic event efficiently
-            let winning_stake = Self::get_outcome_stake(env.clone(), pool_id, outcome);
-
-            PoolResolvedEvent {
-                pool_id,
-                operator,
-                outcome,
-            }
-            .publish(&env);
-
-            PoolResolvedDiagEvent {
-                pool_id,
-                outcome,
-                total_stake: pool.total_stake,
-                winning_stake,
-                timestamp: env.ledger().timestamp(),
-            }
-            .publish(&env);
-        }
-
-        Ok(())
-    }
-
-    /// Mark a pool as ready for resolution and emit an event.
-    /// Can be called by anyone once the resolution delay has passed.
-    pub fn mark_pool_ready(env: Env, pool_id: u64) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-
-        if pool.state != MarketState::Active {
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        let config = Self::get_config(&env);
-        let current_time = env.ledger().timestamp();
-
-        if current_time >= pool.end_time.saturating_add(config.resolution_delay) {
-            PoolReadyForResolutionEvent {
-                pool_id,
-                timestamp: current_time,
-            }
-            .publish(&env);
-            Ok(())
-        } else {
-            Err(PredifiError::ResolutionDelayNotMet)
-        }
-    }
-
-    /// Signal that staking has closed for a pool and emit a `StakingClosedEvent`.
-    ///
-    /// Staking is considered closed once `pool.end_time` has passed — at that
-    /// point no more predictions can be placed.  This function provides an
-    /// explicit, permissionless on-chain signal of that transition so that
-    /// off-chain subscribers (event indexers, front-ends, keepers) can react
-    /// without having to poll block timestamps themselves.
-    ///
-    /// The function is **idempotent**: if it has already been called for the
-    /// given pool, subsequent calls succeed silently without re-emitting the
-    /// event.  This prevents duplicate events even if multiple callers race to
-    /// trigger the transition.
-    ///
-    /// # Arguments
-    /// * `pool_id` - The ID of the pool whose staking window has ended.
-    ///
-    /// # Errors
-    /// - `PoolNotFound`     — no pool exists for `pool_id`.
-    /// - `InvalidPoolState` — pool is not in the `Active` state (already
-    ///                        resolved or cancelled pools have no open staking
-    ///                        window to close).
-    /// - `StakingStillOpen` — `pool.end_time` has not yet been reached; the
-    ///                        staking window is still open.
-    ///
-    /// PRE:  pool.state = Active, env.ledger().timestamp() >= pool.end_time
-    /// POST: StakingClosed(pool_id) sentinel written; StakingClosedEvent emitted
-    ///       exactly once per pool.
-    pub fn close_staking(env: Env, pool_id: u64) -> Result<(), PredifiError> {
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .ok_or(PredifiError::PoolNotFound)?;
-
-        // Only pools that are still Active have a meaningful open staking window.
-        if pool.state != MarketState::Active {
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        let current_time = env.ledger().timestamp();
-
-        // Staking closes when the ledger passes end_time.
-        if current_time < pool.end_time {
-            return Err(PredifiError::StakingStillOpen);
-        }
-
-        // Idempotency guard — if the event has already been emitted for this
-        // pool, return success without re-emitting.
-        let sentinel_key = DataKey::StakingClosed(pool_id);
-        if env.storage().persistent().has(&sentinel_key) {
-            return Ok(());
-        }
-
-        // Mark as closed and emit the event.
-        env.storage().persistent().set(&sentinel_key, &true);
-        Self::extend_persistent(&env, &sentinel_key);
-
-        StakingClosedEvent {
-            pool_id,
-            end_time: pool.end_time,
-            total_stake: pool.total_stake,
-            timestamp: current_time,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Cancel an active pool. Caller must have Operator role (1).
-    /// Cancel a pool, freezing all betting and enabling refund process.
-    /// Only callable by Admin (role 0) - can cancel any pool for any reason.
-    ///
-    /// # Arguments
-    /// * `caller`  - The address requesting the cancellation (must be admin).
-    /// * `pool_id` - The ID of the pool to cancel.
-    /// * `reason`  - A short description of why the pool is being canceled.
-    ///
-    /// # Errors
-    /// - `Unauthorized` if caller is not admin/operator and not the pool creator, or if creator
-    ///   attempts to cancel a pool that already has bets beyond initial liquidity.
-    /// - `PoolNotResolved` error (code 22) is returned if trying to cancel an already resolved pool.
-    /// PRE: pool.state = Active, caller has role 0/1 OR (caller == pool.creator AND total_stake <= initial_liquidity)
-    /// POST: pool.state = Canceled, state transition valid (INV-2)
-    pub fn cancel_pool(
-        env: Env,
-        operator: Address,
-        pool_id: u64,
-        reason: String,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        operator.require_auth();
-
-        // Protect state-modifying external interactions from reentrancy
-        Self::enter_reentrancy_guard(&env);
-
-        let pool_key = DataKey::Pool(pool_id);
-        let mut pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-        Self::extend_persistent(&env, &pool_key);
-
-        // Determine if caller is admin/operator (role 0 or 1)
-        let is_privileged = Self::require_role(&env, &operator, 0).is_ok()
-            || Self::require_role(&env, &operator, 1).is_ok();
-
-        if !is_privileged {
-            // Check if pool is overdue (past end_time + CANCELATION_DELAY)
-            let current_time = env.ledger().timestamp();
-            let overdue_threshold = pool.end_time + CANCELATION_DELAY;
-
-            if current_time > overdue_threshold {
-                // Allow any user to cancel overdue pools
-                // This is a failsafe to unlock funds when resolution is delayed
-            } else {
-                // Allow creator to cancel only if no bets have been placed beyond initial liquidity
-                if operator != pool.creator {
-                    Self::exit_reentrancy_guard(&env);
-                    return Err(PredifiError::Unauthorized);
-                }
-                if pool.total_stake > pool.initial_liquidity {
-                    Self::exit_reentrancy_guard(&env);
-                    return Err(PredifiError::Unauthorized);
-                }
-            }
-        }
-
-        // Ensure only Active pools can be canceled
-        // This prevents canceling pools that are already Resolved or Canceled
-        if !Self::is_pool_active(&pool) {
-            Self::exit_reentrancy_guard(&env);
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        pool.state = MarketState::Canceled;
-        env.storage().persistent().set(&pool_key, &pool);
-        Self::bump_ttl(&env, &pool_key);
-        Self::remove_from_active_index(&env, pool_id);
-
-        PoolCanceledEvent {
-            pool_id,
-            caller: operator.clone(),
-            reason,
-            operator,
-        }
-        .publish(&env);
-
-        Self::exit_reentrancy_guard(&env);
-
-        Ok(())
-    }
-
-    /// Multi-sig emergency cancellation — propose or approve (issue #1119).
-    ///
-    /// Single-signer admin cancellation already exists via `cancel_pool`,
-    /// but for *emergency* cancellations of pools that already hold
-    /// non-trivial stake we want sign-off from more than one privileged
-    /// address. Each call by a distinct admin/operator address records an
-    /// approval; once `EMERGENCY_CANCEL_MULTISIG_THRESHOLD` distinct
-    /// approvals are collected, the pool is moved to `Canceled` exactly
-    /// like the single-signer path.
-    ///
-    /// # Caller
-    /// Any address with role 0 (admin) or 1 (operator).
-    ///
-    /// # Errors
-    /// - `Unauthorized` if caller lacks admin/operator role.
-    /// - `PoolNotFound` if `pool_id` does not exist.
-    /// - `InvalidPoolState` if the pool is not `Active`.
-    /// - `EmergencyCancelAlreadyApproved` if the caller has already approved.
-    /// - `EmergencyCancelPending` is *not* returned — pending state is
-    ///   visible via `get_emergency_cancel_approvals`.
-    pub fn emergency_cancel_pool(
-        env: Env,
-        approver: Address,
-        pool_id: u64,
-        reason: String,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        approver.require_auth();
-
-        // Only admin (0) or operator (1) may participate in emergency cancel.
-        let is_privileged = Self::require_role(&env, &approver, 0).is_ok()
-            || Self::require_role(&env, &approver, 1).is_ok();
-        if !is_privileged {
-            return Err(PredifiError::Unauthorized);
-        }
-
-        Self::enter_reentrancy_guard(&env);
-
-        let pool_key = DataKey::Pool(pool_id);
-        let mut pool: Pool = match env.storage().persistent().get(&pool_key) {
-            Some(p) => p,
-            None => {
-                Self::exit_reentrancy_guard(&env);
-                return Err(PredifiError::PoolNotFound);
-            }
-        };
-        Self::extend_persistent(&env, &pool_key);
-
-        if !Self::is_pool_active(&pool) {
-            Self::exit_reentrancy_guard(&env);
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        // Load existing approvers (or start fresh).
-        let approvers_key = DataKey::EmergencyCancelApprovers(pool_id);
-        let mut approvers: Vec<Address> = env
-            .storage()
-            .persistent()
-            .get(&approvers_key)
-            .unwrap_or_else(|| Vec::new(&env));
-
-        // Reject duplicate approval from the same address.
-        for a in approvers.iter() {
-            if a == approver {
-                Self::exit_reentrancy_guard(&env);
-                return Err(PredifiError::EmergencyCancelAlreadyApproved);
-            }
-        }
-
-        approvers.push_back(approver.clone());
-        env.storage().persistent().set(&approvers_key, &approvers);
-
-        // Capture the first reason; subsequent approvers reaffirm by
-        // approving, but the recorded reason is the proposer's.
-        let reason_key = DataKey::EmergencyCancelReason(pool_id);
-        if !env.storage().persistent().has(&reason_key) {
-            env.storage().persistent().set(&reason_key, &reason);
-        }
-
-        // Below threshold → still pending; surface state but do not cancel.
-        if approvers.len() < EMERGENCY_CANCEL_MULTISIG_THRESHOLD {
-            Self::exit_reentrancy_guard(&env);
-            return Ok(());
-        }
-
-        // Threshold reached — cancel the pool exactly like cancel_pool would.
-        let stored_reason: String = env
-            .storage()
-            .persistent()
-            .get(&reason_key)
-            .unwrap_or(reason);
-
-        pool.state = MarketState::Canceled;
-        env.storage().persistent().set(&pool_key, &pool);
-        Self::bump_ttl(&env, &pool_key);
-        Self::remove_from_active_index(&env, pool_id);
-
-        PoolCanceledEvent {
-            pool_id,
-            caller: approver.clone(),
-            reason: stored_reason,
-            operator: approver,
-        }
-        .publish(&env);
-
-        // Cleanup pending multisig state once the cancel has executed.
-        env.storage().persistent().remove(&approvers_key);
-        env.storage().persistent().remove(&reason_key);
-
-        Self::exit_reentrancy_guard(&env);
-        Ok(())
-    }
-
-    /// Read-only view of pending emergency-cancel approvers for `pool_id`.
-    /// Returns an empty vec when no proposal is currently pending or once
-    /// the threshold has been met and the pool has been cancelled (the
-    /// approvers set is cleared on execution).
-    pub fn get_emergency_cancel_approvals(env: Env, pool_id: u64) -> Vec<Address> {
-        env.storage()
-            .persistent()
-            .get(&DataKey::EmergencyCancelApprovers(pool_id))
-            .unwrap_or_else(|| Vec::new(&env))
-    }
-
-    /// Place a prediction on an active pool by staking tokens on a specific outcome.
-    ///
-    /// This function allows users to participate in prediction markets by transferring tokens
-    /// to the contract and recording their prediction. The prediction is stored and will be
-    /// evaluated when the pool is resolved. Winners can claim their share of the pool's total
-    /// stake minus protocol fees via `claim_winnings`.
-    ///
-    /// # Parameters
-    ///
-    /// - `env` - The Soroban environment, providing access to storage, ledger, and auth
-    /// - `user` - The address placing the prediction (must authenticate via `require_auth`)
-    /// - `pool_id` - The unique identifier of the prediction pool to bet on
-    /// - `amount` - The amount of tokens to stake (must be > 0 and meet minimum requirements)
-    /// - `outcome` - The predicted outcome index (0-based, must be < `pool.options_count`)
-    /// - `referrer` - Optional address that referred this user. If set, the referrer receives
-    ///   a share of the protocol fee when the user claims winnings. Only stored on the first
-    ///   prediction for a given `(user, pool_id)` pair. Cannot be the user or the contract.
-    /// - `invite_key` - Optional symbol used to access private pools. Must match the pool's
-    ///   `whitelist_key` if the pool is private and the user is not whitelisted.
-    ///
-    /// # Prediction Cooldown Mechanism
-    ///
-    /// The contract enforces a cooldown period between consecutive predictions from the same
-    /// address to prevent spam and potential front-running attacks. The cooldown duration is
-    /// configured via `Config::prediction_cooldown_seconds` (default: 0, meaning disabled).
-    ///
-    /// - When `prediction_cooldown_seconds > 0`, the contract checks `LastPredictionTime(user)`
-    ///   to ensure sufficient time has elapsed since the user's last successful prediction.
-    /// - If the cooldown has not elapsed, the function returns `RateLimitOrSuspiciousActivity`.
-    /// - The timestamp is updated after each successful prediction, regardless of pool.
-    ///
-    /// # Stake Limits Enforcement
-    ///
-    /// The function enforces multiple stake limits at different levels:
-    ///
-    /// **Global Protocol Minimum:**
-    /// - `amount` must be >= `Config::min_stake` (default: 1 token unit)
-    /// - Error: `InsufficientStake`
-    ///
-    /// **Per-Pool Limits:**
-    /// - `amount` must be >= `pool.min_stake` (set at pool creation)
-    /// - Error: `StakeBelowMinimum`
-    /// - If `pool.max_stake > 0`, `amount` must be <= `pool.max_stake`
-    /// - Error: `StakeAboveMaximum`
-    ///
-    /// **Total Pool Cap:**
-    /// - If `pool.max_total_stake > 0`, the new `pool.total_stake + amount` must not exceed this cap
-    /// - Error: `MaxTotalStakeExceeded`
-    ///
-    /// **User Prediction Count Limit:**
-    /// - If `Config::max_predictions_per_user > 0`, a user cannot place predictions on more than
-    ///   this number of distinct pools
-    /// - Increasing stake on an existing prediction (same pool, same outcome) does not count
-    ///   toward this limit
-    /// - Error: `MaxPredictionsExceeded`
-    ///
-    /// # Referral Handling
-    ///
-    /// Referrals allow users to earn rewards by bringing new participants to the protocol:
-    ///
-    /// - The `referrer` parameter is only processed on the **first prediction** for a given
-    ///   `(user, pool_id)` pair. Subsequent predictions on the same pool ignore this parameter.
-    /// - The referrer address is stored in `DataKey::Referrer(user, pool_id)`.
-    /// - Referred volume is tracked in `DataKey::ReferredVolume(referrer, pool_id)` and
-    ///   accumulates across all predictions from the referred user on that pool.
-    /// - When the referred user claims winnings, the referrer receives a share of the protocol
-    ///   fee based on `Config::referral_bps` (default: 500 bps = 5%).
-    /// - Referrer validation: Cannot be the user themselves or the contract address.
-    /// - Currently, only one referrer per (user, pool) is supported. Future extensions may
-    ///   support multiple referrers (see `DataKey::Referrer` documentation).
-    ///
-    /// # Fee Deduction Flow
-    ///
-    /// **Note:** Protocol fees are NOT deducted during `place_prediction`. Fees are calculated
-    /// and deducted at **resolution time** when winners claim their winnings via `claim_winnings`.
-    ///
-    /// The fee flow works as follows:
-    ///
-    /// 1. **Prediction Placement (this function):**
-    ///    - User transfers the full `amount` to the contract
-    ///    - The full amount is added to `pool.total_stake`
-    ///    - No fees are deducted at this stage
-    ///
-    /// 2. **Pool Resolution:**
-    ///    - The pool's `fee_bps` is determined by the dynamic fee tier system based on
-    ///      `pool.total_stake` at resolution time
-    ///    - The winning outcome is finalized
-    ///
-    /// 3. **Winnings Claim (`claim_winnings`):**
-    ///    - Protocol fee = `user_winnings * pool.fee_bps / 10_000`
-    ///    - Referral fee = `protocol_fee * Config::referral_bps / 10_000`
-    ///    - Treasury receives: `protocol_fee - referral_fee`
-    ///    - Referrer receives: `referral_fee`
-    ///    - User receives: `user_winnings - protocol_fee`
-    ///
-    /// # Emitted Events
-    ///
-    /// This function emits the following events in order:
-    ///
-    /// 1. **`PredictionPlacedEvent`** (always emitted):
-    ///    - `pool_id` - The pool receiving the prediction
-    ///    - `user` - The address placing the prediction
-    ///    - `amount` - The staked amount
-    ///    - `outcome` - The predicted outcome index
-    ///
-    /// 2. **`HighValuePredictionEvent`** (conditional):
-    ///    - Emitted when `amount >= HIGH_VALUE_THRESHOLD`
-    ///    - Used for monitoring and alerting on large stakes
-    ///    - Fields: `pool_id`, `user`, `amount`, `outcome`, `threshold`
-    ///
-    /// 3. **`OutcomeStakesUpdatedEvent`** (conditional):
-    ///    - Emitted when `pool.options_count >= 16`
-    ///    - Avoids emitting individual events per outcome for large tournaments
-    ///    - Fields: `pool_id`, `options_count`, `total_stake`
-    ///
-    /// 4. **`PredictionBlockedDelistedEvent`** (error case):
-    ///    - Emitted when the pool's token is not whitelisted
-    ///    - Fields: `pool_id`, `user`, `token`, `timestamp`
-    ///
-    /// # Error Conditions
-    ///
-    /// The function can return the following errors:
-    ///
-    /// - `ContractPaused` - The contract is currently paused; all state-mutating operations are blocked
-    /// - `InvalidAmount` - The stake amount is zero or negative
-    /// - `InsufficientStake` - The amount is below the global protocol minimum (`Config::min_stake`)
-    /// - `TokenNotWhitelisted` - The pool's token is not on the allowed betting whitelist
-    /// - `PoolNotFound` - The specified `pool_id` does not exist
-    /// - `InvalidPoolState` - The pool is not in `Active` state (e.g., resolved, canceled, or disputed)
-    /// - `InvalidOutcome` - The outcome index is >= `pool.options_count`
-    /// - `StakeBelowMinimum` - The amount is below the pool's `min_stake`
-    /// - `StakeAboveMaximum` - The amount exceeds the pool's `max_stake` (if > 0)
-    /// - `MaxTotalStakeExceeded` - Adding this amount would exceed `pool.max_total_stake`
-    /// - `RateLimitOrSuspiciousActivity` - The prediction cooldown period has not elapsed
-    /// - `MaxPredictionsExceeded` - The user has exceeded the maximum number of pools they can participate in
-    /// - `ArithmeticError` - An overflow occurred during stake calculations
-    ///
-    /// # Pre-conditions
-    ///
-    /// - `amount > 0` (INV-7)
-    /// - `pool.state == MarketState::Active`
-    /// - `env.ledger().timestamp() < pool.end_time` (pool has not ended)
-    /// - `pool.min_stake <= amount <= pool.max_stake` (unless `max_stake == 0`)
-    /// - `amount >= Config::min_stake` (global minimum)
-    /// - `outcome < pool.options_count`
-    /// - Pool's token must be whitelisted
-    /// - For private pools: user must be whitelisted, be the creator, or provide valid `invite_key`
-    /// - If `prediction_cooldown_seconds > 0`: sufficient time must have elapsed since user's last prediction
-    /// - If `max_predictions_per_user > 0`: user must not have exceeded the prediction count limit
-    ///
-    /// # Post-conditions
-    ///
-    /// - `pool.total_stake` increases by `amount` (INV-1)
-    /// - `OutcomeStake(pool_id, outcome)` increases by `amount` (INV-1)
-    /// - User's prediction record is created or updated with the new stake
-    /// - If first prediction for user on this pool: `participants_count` increments
-    /// - If referrer provided on first prediction: referrer is stored and referred volume is tracked
-    /// - `LastPredictionTime(user)` is updated to current timestamp
-    /// - Tokens are transferred from `user` to the contract
-    /// - `PredictionPlacedEvent` is emitted
-    ///
-    /// # Reentrancy Protection
-    ///
-    /// This function uses a reentrancy guard (`enter_reentrancy_guard` / `exit_reentrancy_guard`)
-    /// to prevent reentrant calls. The guard is entered at the start and exited before token
-    /// transfer and event emission.
-    #[allow(clippy::needless_borrows_for_generic_args)]
-    pub fn place_prediction(
-        env: Env,
-        user: Address,
-        pool_id: u64,
-        amount: i128,
-        outcome: u32,
-        referrer: Option<Address>,
-        invite_key: Option<Symbol>,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        user.require_auth();
-        // Reject zero or negative stake amounts.
-        if amount <= 0 {
-            soroban_sdk::panic_with_error!(&env, PredifiError::InvalidAmount);
-        }
-
-        // Validate: amount must meet the global protocol minimum stake
-        let global_min_stake = Self::get_config(&env).min_stake;
-        if amount < global_min_stake {
-            soroban_sdk::panic_with_error!(&env, PredifiError::InsufficientStake);
-        }
-
-        // Validate referrer if provided: cannot be self or contract
-        if let Some(ref r) = referrer {
-            assert!(r != &user, "referrer cannot be self");
-            assert!(
-                r != &env.current_contract_address(),
-                "referrer cannot be contract"
-            );
-        }
-
-        if let Some(ref invite_key) = invite_key {
-            if let Err(e) = Self::validate_referral_code(&env, invite_key) {
-                soroban_sdk::panic_with_error!(&env, e);
-            }
-        }
-
-        Self::enter_reentrancy_guard(&env);
-
-        let pool_key = DataKey::Pool(pool_id);
-        let mut pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-
-        // assert!(pool.state == MarketState::Active, "Pool is not active");
-        if !Self::is_pool_active(&pool) {
-            panic!("Pool is not active");
-        }
-        assert!(env.ledger().timestamp() < pool.end_time, "Pool has ended");
-
-        // Validate: token must be on the allowed betting whitelist
-        if !Self::is_token_whitelisted(&env, &pool.token) {
-            Self::exit_reentrancy_guard(&env);
-            PredictionBlockedDelistedEvent {
-                pool_id,
-                user: user.clone(),
-                token: pool.token.clone(),
-                timestamp: env.ledger().timestamp(),
-            }
-            .publish(&env);
-            soroban_sdk::panic_with_error!(&env, PredifiError::TokenNotWhitelisted);
-        }
-
-        // Check private pool authorization
-        // Check private pool authorization
-        if pool.private {
-            let whitelist_key_data = DataKey::Whitelist(pool_id, user.clone());
-            let is_whitelisted = env
-                .storage()
-                .persistent()
-                .get(&whitelist_key_data)
-                .unwrap_or(false);
-
-            let has_valid_invite = if let Some(ref pool_key) = pool.whitelist_key {
-                if let Some(ref prov_key) = invite_key {
-                    pool_key == prov_key
-                } else {
-                    false
-                }
-            } else {
-                false
-            };
-
-            assert!(
-                is_whitelisted || user == pool.creator || has_valid_invite,
-                "User not authorized for private pool"
-            );
-        }
-
-        // Validate: outcome must be within the valid options range
-        if outcome >= pool.options_count {
-            soroban_sdk::panic_with_error!(&env, PredifiError::InvalidOutcome);
-        }
-
-        // --- INTERNAL CHECKS & EFFECTS ---
-        // Validate: per-pool stake limits
-        if amount < pool.min_stake {
-            Self::exit_reentrancy_guard(&env);
-            soroban_sdk::panic_with_error!(&env, PredifiError::StakeBelowMinimum);
-        }
-        if pool.max_stake > 0 && amount > pool.max_stake {
-            Self::exit_reentrancy_guard(&env);
-            soroban_sdk::panic_with_error!(&env, PredifiError::StakeAboveMaximum);
-        }
-
-        // Enforce global pool cap (max total stake)
-        if pool.max_total_stake > 0 {
-            let new_total = pool.total_stake.checked_add(amount).expect("overflow");
-            if new_total > pool.max_total_stake {
-                Self::exit_reentrancy_guard(&env);
-                soroban_sdk::panic_with_error!(&env, PredifiError::MaxTotalStakeExceeded);
-            }
-        }
-
-        // Enforce maximum predictions per user limit (across all pools)
-        let config = Self::get_config(&env);
-        if config.prediction_cooldown_seconds > 0 {
-            let last_prediction_key = DataKey::LastPredictionTime(user.clone());
-            if let Some(last_prediction_time) = env
-                .storage()
-                .persistent()
-                .get::<_, u64>(&last_prediction_key)
-            {
-                Self::extend_persistent(&env, &last_prediction_key);
-                let now = env.ledger().timestamp();
-                if now.saturating_sub(last_prediction_time) < config.prediction_cooldown_seconds {
-                    Self::exit_reentrancy_guard(&env);
-                    soroban_sdk::panic_with_error!(
-                        &env,
-                        PredifiError::RateLimitOrSuspiciousActivity
-                    );
-                }
-            }
-        }
-
-        if config.max_predictions_per_user > 0 {
-            let pred_key = DataKey::Pred(user.clone(), pool_id);
-            let existing_pred = env.storage().persistent().get::<_, Prediction>(&pred_key);
-
-            // If user already has a prediction on this pool, allow increasing stake (same prediction)
-            // If this is a new prediction for this pool, check if user has reached the limit
-            if existing_pred.is_none() {
-                // Count current number of pools this user has predictions in
-                let user_prediction_count_key = DataKey::UsrPrdCnt(user.clone());
-                let current_count: u32 = env
-                    .storage()
-                    .persistent()
-                    .get(&user_prediction_count_key)
-                    .unwrap_or(0);
-
-                if current_count >= config.max_predictions_per_user {
-                    Self::exit_reentrancy_guard(&env);
-                    soroban_sdk::panic_with_error!(&env, PredifiError::MaxPredictionsExceeded);
-                }
-            }
-            // Note: If user already has a prediction on this pool, we allow increasing the stake
-            // as it's the same prediction, not a new pool participation
-        }
-
-        let pred_key = DataKey::Pred(user.clone(), pool_id);
-        let existing_pred = env.storage().persistent().get::<_, Prediction>(&pred_key);
-        if let Some(mut existing_pred) = existing_pred {
-            assert!(
-                existing_pred.outcome == outcome,
-                "Cannot change prediction outcome"
-            );
-            existing_pred.amount = existing_pred.amount.checked_add(amount).expect("overflow");
-            env.storage().persistent().set(&pred_key, &existing_pred);
-            Self::extend_persistent(&env, &pred_key);
-
-            // Track referred volume: if this user already has a referrer, add to their volume
-            let referrer_key = DataKey::Referrer(user.clone(), pool_id);
-            if let Some(referrer_addr) = env.storage().persistent().get::<_, Address>(&referrer_key)
-            {
-                Self::extend_persistent(&env, &referrer_key);
-                let vol_key = DataKey::ReferredVolume(referrer_addr.clone(), pool_id);
-                let vol: i128 = env.storage().persistent().get(&vol_key).unwrap_or(0);
-                env.storage().persistent().set(&vol_key, &(vol + amount));
-                Self::extend_persistent(&env, &vol_key);
-            }
-        } else {
-            env.storage()
-                .persistent()
-                .set(&pred_key, &Prediction { amount, outcome });
-            Self::extend_persistent(&env, &pred_key);
-
-            // Store referrer on first prediction and track referred volume.
-            // NOTE: Only one referrer per (user, pool) is supported today.
-            // See DataKey::Referrer for a note on extending this to multiple referrers.
-            if let Some(ref referrer_addr) = referrer {
-                let referrer_key = DataKey::Referrer(user.clone(), pool_id);
-                env.storage().persistent().set(&referrer_key, referrer_addr);
-                Self::extend_persistent(&env, &referrer_key);
-                let vol_key = DataKey::ReferredVolume(referrer_addr.clone(), pool_id);
-                let vol: i128 = env.storage().persistent().get(&vol_key).unwrap_or(0);
-                env.storage().persistent().set(&vol_key, &(vol + amount));
-                Self::extend_persistent(&env, &vol_key);
-            }
-
-            // Increment participants_count in the pool struct
-            pool.participants_count = pool.participants_count.saturating_add(1);
-
-            let count_key = DataKey::UsrPrdCnt(user.clone());
-            let count: u32 = env.storage().persistent().get(&count_key).unwrap_or(0);
-
-            let index_key = DataKey::UsrPrdIdx(user.clone(), count);
-            env.storage().persistent().set(&index_key, &pool_id);
-            Self::extend_persistent(&env, &index_key);
-
-            env.storage().persistent().set(&count_key, &(count + 1));
-            Self::extend_persistent(&env, &count_key);
-        }
-
-        // Update total stake (INV-1)
-        pool.total_stake = pool.total_stake.checked_add(amount).expect("overflow");
-        env.storage().persistent().set(&pool_key, &pool);
-        Self::bump_ttl(&env, &pool_key);
-
-        let last_prediction_key = DataKey::LastPredictionTime(user.clone());
-        env.storage()
-            .persistent()
-            .set(&last_prediction_key, &env.ledger().timestamp());
-        Self::extend_persistent(&env, &last_prediction_key);
-
-        // Update outcome stake (INV-1) - using optimized batch storage
-        let _stakes =
-            Self::update_outcome_stake(&env, pool_id, outcome, amount, pool.options_count);
-
-        // --- INTERACTIONS ---
-
-        // Validate token transfer safety before executing
-        Self::validate_token_transfer(
-            &env,
-            &pool.token,
-            &user,
-            &env.current_contract_address(),
-            amount,
-        )?;
-
-        let token_client = token::Client::new(&env, &pool.token);
-        token_client.transfer(&user, &env.current_contract_address(), &amount);
-
-        Self::exit_reentrancy_guard(&env);
-
-        PredictionPlacedEvent {
-            pool_id,
-            user: user.clone(),
-            amount,
-            outcome,
-        }
-        .publish(&env);
-
-        // 🟡 MEDIUM ALERT: large stake detected — emit supplementary event.
-        if amount >= HIGH_VALUE_THRESHOLD {
-            HighValuePredictionEvent {
-                pool_id,
-                user,
-                amount,
-                outcome,
-                threshold: HIGH_VALUE_THRESHOLD,
-            }
-            .publish(&env);
-        }
-
-        // 🟢 INFO: For markets with many outcomes (16+), emit batch stake update event
-        // to avoid emitting individual events per outcome which would be impractical
-        // for large tournaments (e.g., 32-team bracket).
-        if pool.options_count >= 16 {
-            OutcomeStakesUpdatedEvent {
-                pool_id,
-                options_count: pool.options_count,
-                total_stake: pool.total_stake,
-            }
-            .publish(&env);
-        }
-
-        Ok(())
-    }
-
-    /// Claim winnings from a resolved pool. Returns the amount paid out (0 for losers).
-    /// PRE: pool.state ≠ Active
-    /// POST: HasClaimed(user, pool) = true (INV-3), payout ≤ pool.total_stake (INV-4)
-    fn claim_winnings_internal(
-        env: &Env,
-        user: &Address,
-        pool_id: u64,
-    ) -> Result<i128, PredifiError> {
-        Self::enter_reentrancy_guard(env);
-
-        let result: Result<i128, PredifiError> = (|| {
-            let pool_key = DataKey::Pool(pool_id);
-            let pool: Pool = env
-                .storage()
-                .persistent()
-                .get(&pool_key)
-                .expect("Pool not found");
-            Self::extend_persistent(env, &pool_key);
-
-            if pool.state == MarketState::Active {
-                return Err(PredifiError::PoolNotResolved);
-            }
-
-            let claimed_key = DataKey::Claimed(user.clone(), pool_id);
-            if env.storage().persistent().has(&claimed_key) {
-                SuspiciousDoubleClaimEvent {
-                    user: user.clone(),
-                    pool_id,
-                    timestamp: env.ledger().timestamp(),
-                }
-                .publish(env);
-                return Err(PredifiError::AlreadyClaimed);
-            }
-
-            let pred_key = DataKey::Pred(user.clone(), pool_id);
-            // Single get — avoid redundant has() storage read on the hot claim path
-            let prediction: Option<Prediction> = env.storage().persistent().get(&pred_key);
-            let prediction = match prediction {
-                Some(p) => {
-                    Self::extend_persistent(env, &pred_key);
-                    p
-                }
-                None => return Ok(0),
-            };
-
-            env.storage().persistent().set(&claimed_key, &true);
-            Self::bump_ttl(env, &claimed_key);
-
-            if pool.state == MarketState::Canceled {
-                // Validate token transfer before sending refund
-                Self::validate_token_transfer(
-                    env,
-                    &pool.token,
-                    &env.current_contract_address(),
-                    user,
-                    prediction.amount,
-                )?;
-
-                let token_client = token::Client::new(env, &pool.token);
-                token_client.transfer(&env.current_contract_address(), user, &prediction.amount);
-
-                WinningsClaimedEvent {
-                    pool_id,
-                    user: user.clone(),
-                    amount: prediction.amount,
-                }
-                .publish(env);
-
-                RewardClaimedEvent {
-                    pool_id,
-                    user: user.clone(),
-                    amount: prediction.amount,
-                    claim_type: String::from_str(env, "winnings"),
-                }
-                .publish(env);
-
-                return Ok(prediction.amount);
-            }
-
-            // Check if pool is properly resolved
-            if !Self::is_pool_resolved(&pool) {
-                return Err(PredifiError::PoolNotResolved);
-            }
-
-            // Check claim window expiration if resolution timestamp exists
-            if let Some(resolution_timestamp) = pool.resolution_timestamp {
-                let config = Self::get_config(env);
-                let claim_deadline = resolution_timestamp
-                    .checked_add(config.claim_window_seconds)
-                    .ok_or(PredifiError::InvalidTimestamp)?;
-                let current_time = env.ledger().timestamp();
-
-                if current_time > claim_deadline {
-                    return Err(PredifiError::InvalidTimestamp);
-                }
-            }
-
-            if prediction.outcome != pool.outcome {
-                return Ok(0);
-            }
-
-            let winning_stake = Self::get_outcome_stake(env.clone(), pool_id, pool.outcome);
-
-            if winning_stake == 0 {
-                return Ok(0);
-            }
-
-            let fee_bps_i = if pool.fee_bps > 0 || pool.state == MarketState::Resolved {
-                pool.fee_bps as i128
-            } else {
-                let config = Self::get_config(env);
-                config.fee_bps as i128
-            };
-
-            // Payout math lives in `payouts` — keeps lib.rs focused on orchestration
-            let breakdown = calculate_claim_payout(&PayoutInput {
-                pool_total_stake: pool.total_stake,
-                fee_bps: fee_bps_i,
-                user_stake: prediction.amount,
-                winning_stake,
-            })
-            .map_err(|_| PredifiError::InvalidAmount)?;
-            let protocol_fee_total = breakdown.protocol_fee;
-            let winnings = breakdown.winnings;
-
-            assert!(winnings <= pool.total_stake, "Winnings exceed total stake");
-
-            let token_client = token::Client::new(env, &pool.token);
-
-            let referrer_key = DataKey::Referrer(user.clone(), pool_id);
-            if let Some(referrer) = env.storage().persistent().get::<_, Address>(&referrer_key) {
-                Self::extend_persistent(env, &referrer_key);
-                if protocol_fee_total > 0 && pool.total_stake > 0 {
-                    let referral_cut_bps = Self::read_referral_cut_bps(env) as i128;
-                    let referral_amount = calculate_referral_amount(
-                        prediction.amount,
-                        pool.total_stake,
-                        protocol_fee_total,
-                        referral_cut_bps,
-                    )
-                    .map_err(|_| PredifiError::InvalidAmount)?;
-                    if referral_amount > 0 {
-                        // Validate referral token transfer before execution
-                        Self::validate_token_transfer(
-                            env,
-                            &pool.token,
-                            &env.current_contract_address(),
-                            &referrer,
-                            referral_amount,
-                        )?;
-
-                        token_client.transfer(
-                            &env.current_contract_address(),
-                            &referrer,
-                            &referral_amount,
-                        );
-                        ReferralPaidEvent {
-                            pool_id,
-                            referrer: referrer.clone(),
-                            referred_user: user.clone(),
-                            amount: referral_amount,
-                        }
-                        .publish(env);
-                    }
-                }
-            }
-
-            if winnings > 0 {
-                // Validate main winnings transfer before execution
-                Self::validate_token_transfer(
-                    env,
-                    &pool.token,
-                    &env.current_contract_address(),
-                    user,
-                    winnings,
-                )?;
-
-                token_client.transfer(&env.current_contract_address(), user, &winnings);
-            }
-
-            WinningsClaimedEvent {
-                pool_id,
-                user: user.clone(),
-                amount: winnings,
-            }
-            .publish(env);
-
-            RewardClaimedEvent {
-                pool_id,
-                user: user.clone(),
-                amount: winnings,
-                claim_type: String::from_str(env, "winnings"),
-            }
-            .publish(env);
-
-            Ok(winnings)
-        })();
-
-        Self::exit_reentrancy_guard(env);
-        result
-    }
-
-    /// Claim winning payout from a resolved prediction market pool.
-    ///
-    /// # Payout Calculation & Economics
-    /// - Payout is calculated proportionally based on the caller's stake relative to total winning stakes:
-    ///   `user_payout = (user_stake / winning_stake) * (total_pool_stake - protocol_fee)`
-    /// - Protocol fee is deducted based on the pool's configured basis points (`fee_bps`).
-    /// - If a referrer is associated with the user, a portion of the fee is transferred to the referrer.
-    ///
-    /// # Claim Window Enforcement
-    /// - If `claim_window_seconds` is configured, claims are strictly rejected with `InvalidTimestamp`
-    ///   if the current ledger timestamp exceeds `resolution_timestamp + claim_window_seconds`.
-    ///
-    /// # Double-Claim Prevention & Security
-    /// - Prevents double-claiming by storing a `DataKey::Claimed(user, pool_id)` sentinel in persistent storage.
-    /// - Re-entrancy guard (`enter_reentrancy_guard` / `exit_reentrancy_guard`) protects token transfers.
-    /// - Suspicious double-claim attempts trigger `SuspiciousDoubleClaimEvent` alerts.
-    ///
-    /// # Token Transfer Mechanics
-    /// - Validates token transfer limits via `validate_token_transfer`.
-    /// - Executes token transfer directly from the contract address to the user's address via Soroban `token::Client`.
-    ///
-    /// # Emitted Events
-    /// - Emits `WinningsClaimedEvent` and `RewardClaimedEvent` upon successful claim.
-    /// - Emits `ReferralPaidEvent` if a referral reward is distributed.
-    ///
-    /// # Arguments
-    /// * `env` - The Soroban host environment.
-    /// * `user` - Address of the winning user claiming payout (must require_auth).
-    /// * `pool_id` - ID of the resolved market pool.
-    ///
-    /// # Returns
-    /// * `Ok(amount)` - Net winning token amount transferred to `user`.
-    /// * `Err(PredifiError)` - Reason for claim failure (e.g. `PoolNotResolved`, `AlreadyClaimed`, `InvalidTimestamp`).
-    #[allow(clippy::needless_borrows_for_generic_args)]
-    pub fn claim_winnings(env: Env, user: Address, pool_id: u64) -> Result<i128, PredifiError> {
-        Self::require_not_paused(&env)?;
-        user.require_auth();
-        Self::claim_winnings_internal(&env, &user, pool_id)
-    }
-
-    /// Claim winnings from multiple pools in a single transaction.
-    ///
-    /// Iterates over `pool_ids`, calls the single-pool claim logic for each,
-    /// and returns a `Map<u64, i128>` showing how much was claimed per pool.
-    /// Pools that yield 0 (loser, already claimed, no prediction) are still
-    /// included in the map with value 0 so callers can distinguish "processed"
-    /// from "not attempted".
-    ///
-    /// # Arguments
-    /// * `user`     - Address claiming winnings (must provide auth once)
-    /// * `pool_ids` - List of pool IDs to claim from
-    ///
-    /// # Returns
-    /// `Map<u64, i128>` — claimed amount per pool (0 for non-winners / already claimed)
-    pub fn batch_claim_winnings(
-        env: Env,
-        user: Address,
-        pool_ids: Vec<u64>,
-    ) -> Result<soroban_sdk::Map<u64, i128>, PredifiError> {
-        Self::require_not_paused(&env)?;
-        user.require_auth();
-        let mut results: soroban_sdk::Map<u64, i128> = soroban_sdk::Map::new(&env);
-        for pool_id in pool_ids.iter() {
-            let amount = Self::claim_winnings_internal(&env, &user, pool_id).unwrap_or(0);
-            results.set(pool_id, amount);
-        }
-        Ok(results)
-    }
-
-    /// Claim a refund from a canceled prediction pool.
-    ///
-    /// # Refund Calculation & Economics
-    /// - Returns 100% of the user's original staked principal amount (`prediction.amount`).
-    /// - No protocol fees or penalties are deducted when a market pool is canceled.
-    ///
-    /// # Double-Claim Prevention & Security
-    /// - Enforces `INV-3` double-claim prevention by writing `DataKey::Claimed(user, pool_id)` to persistent storage.
-    /// - Protected by re-entrancy guard (`enter_reentrancy_guard` / `exit_reentrancy_guard`) during asset transfers.
-    ///
-    /// # Token Transfer Mechanics
-    /// - Checks balance and validates transfer via `validate_token_transfer`.
-    /// - Transfers funds from contract balance directly to `user` via Soroban `token::Client`.
-    ///
-    /// # Emitted Events
-    /// - Emits `RefundClaimedEvent` with `pool_id`, `user`, and refunded `amount`.
-    /// - Emits `RewardClaimedEvent` with `claim_type: "refund"`.
-    ///
-    /// PRE: pool.state = Canceled, user has an active prediction on the pool.
-    /// POST: HasClaimed(user, pool) = true (INV-3), user receives full principal stake amount.
-    ///
-    /// # Arguments
-    /// * `env` - The Soroban host environment instance.
-    /// * `user` - Address claiming the refund (must provide authorization).
-    /// * `pool_id` - Unique identifier of the canceled pool.
-    ///
-    /// # Returns
-    /// * `Ok(amount)` - Refund successfully claimed, returns exact refunded principal amount.
-    /// * `Err(PredifiError)` - Operation failed with specific error code.
-    ///
-    /// # Errors
-    /// - `InvalidPoolState` if pool doesn't exist or is not in `Canceled` state.
-    /// - `InsufficientBalance` if user has no stake to refund.
-    /// - `AlreadyClaimed` if user has already claimed a refund for this pool.
-    #[allow(clippy::needless_borrows_for_generic_args)]
-    pub fn claim_refund(env: Env, user: Address, pool_id: u64) -> Result<i128, PredifiError> {
-        Self::require_not_paused(&env)?;
-        user.require_auth();
-
-        // 🛡️ RE-ENTRANCY GUARD: Protect against recursive withdrawal attempts
-        // during value transfer to external addresses/contracts (INV-3).
-        Self::enter_reentrancy_guard(&env);
-
-        let result: Result<i128, PredifiError> = (|| {
-            // --- CHECKS ---
-
-            let pool_key = DataKey::Pool(pool_id);
-            let pool: Pool = match env.storage().persistent().get(&pool_key) {
-                Some(p) => p,
-                None => {
-                    return Err(PredifiError::InvalidPoolState);
-                }
-            };
-            Self::extend_persistent(&env, &pool_key);
-
-            // Verify pool is canceled
-            if pool.state != MarketState::Canceled {
-                return Err(PredifiError::InvalidPoolState);
-            }
-
-            // Check if user already claimed refund
-            let claimed_key = DataKey::Claimed(user.clone(), pool_id);
-            if env.storage().persistent().has(&claimed_key) {
-                return Err(PredifiError::AlreadyClaimed);
-            }
-
-            // Get user's prediction
-            let pred_key = DataKey::Pred(user.clone(), pool_id);
-            let prediction: Option<Prediction> = env.storage().persistent().get(&pred_key);
-
-            if env.storage().persistent().has(&pred_key) {
-                Self::extend_persistent(&env, &pred_key);
-            }
-
-            let prediction = match prediction {
-                Some(p) => p,
-                None => {
-                    return Err(PredifiError::InsufficientBalance);
-                }
-            };
-
-            // Verify user has a non-zero stake
-            if prediction.amount <= 0 {
-                return Err(PredifiError::InsufficientBalance);
-            }
-
-            // --- EFFECTS ---
-
-            // Mark as claimed immediately to prevent re-entrancy (INV-3)
-            env.storage().persistent().set(&claimed_key, &true);
-            Self::bump_ttl(&env, &claimed_key);
-
-            let refund_amount = prediction.amount;
-
-            // --- INTERACTIONS ---
-
-            // Validate token transfer before sending refund
-            Self::validate_token_transfer(
-                &env,
-                &pool.token,
-                &env.current_contract_address(),
-                &user,
-                refund_amount,
-            )?;
-
-            let token_client = token::Client::new(&env, &pool.token);
-            token_client.transfer(&env.current_contract_address(), &user, &refund_amount);
-
-            RefundClaimedEvent {
-                pool_id,
-                user: user.clone(),
-                amount: refund_amount,
-            }
-            .publish(&env);
-
-            RewardClaimedEvent {
-                pool_id,
-                user: user.clone(),
-                amount: refund_amount,
-                claim_type: String::from_str(&env, "refund"),
-            }
-            .publish(&env);
-
-            Ok(refund_amount)
-        })();
-
-        Self::exit_reentrancy_guard(&env);
-        result
-    }
-
-    /// Update the stake limits for an active pool. Caller must have Operator role (1).
-    /// PRE: pool.state = Active, operator has role 1
-    /// POST: pool.min_stake and pool.max_stake updated
-    pub fn set_stake_limits(
-        env: Env,
-        operator: Address,
-        pool_id: u64,
-        min_stake: i128,
-        max_stake: i128,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        operator.require_auth();
-        Self::require_role(&env, &operator, 1)?;
-
-        let pool_key = DataKey::Pool(pool_id);
-        let mut pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-
-        if pool.state != MarketState::Active {
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        // Validate new stake limits before applying
-        Self::validate_stake_limits(&env, &pool, min_stake, max_stake)?;
-
-        pool.min_stake = min_stake;
-        pool.max_stake = max_stake;
-
-        env.storage().persistent().set(&pool_key, &pool);
-        Self::extend_persistent(&env, &pool_key);
-
-        StakeLimitsUpdatedEvent {
-            pool_id,
-            operator,
-            min_stake,
-            max_stake,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Get a paginated list of a user's predictions.
-    ///
-    /// # Errors
-    /// Returns `PredifiError::InvalidPagination` if `offset + limit` overflows `u32`.
-    pub fn get_user_predictions(
-        env: Env,
-        user: Address,
-        offset: u32,
-        limit: u32,
-    ) -> Result<Vec<UserPredictionDetail>, PredifiError> {
-        // Guard against offset + limit wrapping around u32::MAX.
-        let end_check = offset
-            .checked_add(limit)
-            .ok_or(PredifiError::InvalidPagination)?;
-
-        let count_key = DataKey::UsrPrdCnt(user.clone());
-        let count: u32 = env.storage().persistent().get(&count_key).unwrap_or(0);
-        if env.storage().persistent().has(&count_key) {
-            Self::extend_persistent(&env, &count_key);
-        }
-
-        let mut results = Vec::new(&env);
-
-        if offset >= count || limit == 0 {
-            return Ok(results);
-        }
-
-        let end = core::cmp::min(end_check, count);
-
-        for i in offset..end {
-            let index_key = DataKey::UsrPrdIdx(user.clone(), i);
-            let pool_id: u64 = env
-                .storage()
-                .persistent()
-                .get(&index_key)
-                .expect("index not found");
-            Self::extend_persistent(&env, &index_key);
-
-            let pred_key = DataKey::Pred(user.clone(), pool_id);
-            let prediction: Prediction = env
-                .storage()
-                .persistent()
-                .get(&pred_key)
-                .expect("prediction not found");
-            Self::extend_persistent(&env, &pred_key);
-
-            let pool_key = DataKey::Pool(pool_id);
-            let pool: Pool = env
-                .storage()
-                .persistent()
-                .get(&pool_key)
-                .expect("pool not found");
-            Self::extend_persistent(&env, &pool_key);
-
-            results.push_back(UserPredictionDetail {
-                pool_id,
-                amount: prediction.amount,
-                user_outcome: prediction.outcome,
-                pool_end_time: pool.end_time,
-                pool_state: pool.state,
-                pool_outcome: pool.outcome,
-            });
-        }
-
-        Ok(results)
-    }
-
-    /// This function is optimized for markets with many outcomes (e.g., 32+ teams).
-    /// Instead of making N storage reads (one per outcome), it makes a single read.
-    ///
-    /// Returns a Vec of stakes where index corresponds to outcome index.
-    /// For example, `stake\[0\]` is the total amount bet on outcome 0.
-    pub fn get_pool(env: Env, pool_id: u64) -> Pool {
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-        Self::extend_persistent(&env, &pool_key);
-        pool
-    }
-
-    /// Returns the configuration fields of a pool as a `PoolConfig` struct.
-    ///
-    /// This is a lightweight alternative to `get_pool` when only the
-    /// configuration parameters are needed (description, stake limits, etc.)
-    /// without fetching the full runtime state (total_stake, outcome, etc.).
-    ///
-    /// # Panics
-    /// Panics with "Pool not found" if no pool exists for the given `pool_id`.
-    pub fn get_pool_config(env: Env, pool_id: u64) -> PoolConfig {
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-        Self::extend_persistent(&env, &pool_key);
-        PoolConfig {
-            start_time: pool.start_time,
-            description: pool.description,
-            metadata_url: pool.metadata_url,
-            min_stake: pool.min_stake,
-            max_stake: pool.max_stake,
-            min_total_stake: pool.min_total_stake,
-            max_total_stake: pool.max_total_stake,
-            initial_liquidity: pool.initial_liquidity,
-            required_resolutions: pool.required_resolutions,
-            private: pool.private,
-            whitelist_key: pool.whitelist_key,
-            outcome_descriptions: pool.outcome_descriptions,
-        }
-    }
-
-    pub fn get_pool_outcome_stakes(env: Env, pool_id: u64) -> Vec<i128> {
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-        Self::extend_persistent(&env, &pool_key);
-
-        Self::get_outcome_stakes(&env, pool_id, pool.options_count)
-    }
-
-    /// Get a specific outcome's stake (backward compatible).
-    ///
-    /// Prefers the batch `OutStakes` key (single read for all outcomes). Falls
-    /// back to the legacy per-outcome `OutStake` key for pre-migration data.
-    pub fn get_outcome_stake(env: Env, pool_id: u64, outcome: u32) -> i128 {
-        // Prefer batch key — canonical after gas optimization (no dual-write)
-        let batch_key = DataKey::OutStakes(pool_id);
-        if let Some(stakes) = env.storage().persistent().get::<_, Vec<i128>>(&batch_key) {
-            Self::extend_persistent(&env, &batch_key);
-            return stakes.get(outcome).unwrap_or(0);
-        }
-
-        // Legacy fallback: individual key
-        let stake_key = DataKey::OutStake(pool_id, outcome);
-        if let Some(stake) = env.storage().persistent().get::<_, i128>(&stake_key) {
-            Self::extend_persistent(&env, &stake_key);
-            return stake;
-        }
-
-        0
-    }
-
-    /// Get a paginated list of pool IDs by category.
-    ///
-    /// # Errors
-    /// Returns `PredifiError::InvalidPagination` if `offset + limit` overflows `u32`.
-    pub fn get_pools_by_category(
-        env: Env,
-        category: Symbol,
-        offset: u32,
-        limit: u32,
-    ) -> Result<Vec<u64>, PredifiError> {
-        // Guard against offset + limit wrapping around u32::MAX.
-        offset
-            .checked_add(limit)
-            .ok_or(PredifiError::InvalidPagination)?;
-
-        let count_key = DataKey::CatPoolCt(category.clone());
-        let count: u32 = if let Some(c) = env.storage().persistent().get(&count_key) {
-            Self::extend_persistent(&env, &count_key);
-            c
-        } else {
-            0
-        };
-
-        let mut results = Vec::new(&env);
-
-        if offset >= count || limit == 0 {
-            return Ok(results);
-        }
-
-        let start_index = count.saturating_sub(offset).saturating_sub(1);
-        let num_to_take = core::cmp::min(limit, count.saturating_sub(offset));
-
-        for i in 0..num_to_take {
-            let index = start_index.saturating_sub(i);
-            let index_key = DataKey::CatPoolIx(category.clone(), index);
-            let pool_id: u64 = env
-                .storage()
-                .persistent()
-                .get(&index_key)
-                .expect("index not found");
-            Self::extend_persistent(&env, &index_key);
-
-            results.push_back(pool_id);
-        }
-
-        Ok(results)
-    }
-
-    /// Get a paginated list of all currently active pool IDs across all categories.
-    ///
-    /// Returns pool IDs in insertion order (oldest first within each page).
-    /// Pools are removed from this list when they are resolved or canceled,
-    /// so every ID returned is guaranteed to belong to an active pool.
-    ///
-    /// # Arguments
-    /// * `offset` - Number of entries to skip (0-based).
-    /// * `limit`  - Maximum number of entries to return.
-    ///
-    /// # Returns
-    /// A `Vec<u64>` of active pool IDs. Returns an empty vec if `offset`
-    /// is beyond the current count or `limit` is 0.
-    /// # Errors
-    /// Returns `PredifiError::InvalidPagination` if `offset + limit` overflows `u32`.
-    pub fn get_active_pools(env: Env, offset: u32, limit: u32) -> Result<Vec<u64>, PredifiError> {
-        // Guard against offset + limit wrapping around u32::MAX.
-        let end_check = offset
-            .checked_add(limit)
-            .ok_or(PredifiError::InvalidPagination)?;
-
-        let ctr_key = DataKey::ActivePoolCtr;
-        let count: u32 = env.storage().persistent().get(&ctr_key).unwrap_or(0);
-        let mut results = Vec::new(&env);
-
-        if offset >= count || limit == 0 {
-            return Ok(results);
-        }
-
-        // Only extend the counter TTL if the key actually exists.
-        if count > 0 {
-            Self::extend_persistent(&env, &ctr_key);
-        }
-
-        let end = core::cmp::min(end_check, count);
-
-        for i in offset..end {
-            let slot_key = DataKey::ActivePool(i);
-            if let Some(pool_id) = env.storage().persistent().get(&slot_key) {
-                Self::extend_persistent(&env, &slot_key);
-                results.push_back(pool_id);
-            }
-        }
-
-        Ok(results)
-    }
-
-    /// Return the total number of currently active (open) pools.
-    ///
-    /// This is an O(1) read of the `ActivePoolCtr` persistent-storage counter
-    /// that is maintained by `add_to_active_index` / `remove_from_active_index`.
-    /// Frontends can use this to display "Showing N of M active pools" without
-    /// fetching every page.
-    pub fn get_active_pools_count(env: Env) -> u32 {
-        env.storage()
-            .persistent()
-            .get(&DataKey::ActivePoolCtr)
-            .unwrap_or(0)
-    }
-
-    /// Add a user to a private pool's whitelist. Only callable by pool creator.
-    pub fn add_to_whitelist(
-        env: Env,
-        creator: Address,
-        pool_id: u64,
-        user: Address,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        creator.require_auth();
-
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-        Self::extend_persistent(&env, &pool_key);
-
-        if pool.creator != creator {
-            return Err(PredifiError::Unauthorized);
-        }
-
-        assert!(pool.private, "Pool is not private");
-
-        let whitelist_key = DataKey::Whitelist(pool_id, user.clone());
-        env.storage().persistent().set(&whitelist_key, &true);
-        Self::extend_persistent(&env, &whitelist_key);
-
-        AddedToWhitelistEvent {
-            pool_id,
-            user,
-            added_by: creator,
-            timestamp: env.ledger().timestamp(),
-        }
-        .publish(&env);
-        Ok(())
-    }
-
-    /// Remove a user from a private pool's whitelist. Only callable by pool creator.
-    pub fn remove_from_whitelist(
-        env: Env,
-        creator: Address,
-        pool_id: u64,
-        user: Address,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        creator.require_auth();
-
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-        Self::extend_persistent(&env, &pool_key);
-
-        if pool.creator != creator {
-            return Err(PredifiError::Unauthorized);
-        }
-
-        assert!(pool.private, "Pool is not private");
-
-        let whitelist_key = DataKey::Whitelist(pool_id, user.clone());
-        env.storage().persistent().remove(&whitelist_key);
-
-        RemovedFromWhitelistEvent {
-            pool_id,
-            user,
-            removed_by: creator,
-            timestamp: env.ledger().timestamp(),
-        }
-        .publish(&env);
-        Ok(())
-    }
-
-    /// Check whether a user has an explicit whitelist entry for a pool.
-    ///
-    /// This helper only reports stored whitelist membership. It does not treat
-    /// public pools, pool creators, or invite-based access as implicit
-    /// whitelist membership.
-    pub fn is_whitelisted(env: Env, pool_id: u64, user: Address) -> bool {
-        let whitelist_key = DataKey::Whitelist(pool_id, user);
-        let is_whitelisted = env
-            .storage()
-            .persistent()
-            .get(&whitelist_key)
-            .unwrap_or(false);
-        if env.storage().persistent().has(&whitelist_key) {
-            Self::extend_persistent(&env, &whitelist_key);
-        }
-        is_whitelisted
-    }
-
-    /// Batch add multiple users to a private pool's whitelist.
-    ///
-    /// This optimized function reduces storage operations by processing
-    /// multiple users in a single transaction, improving gas efficiency.
-    ///
-    /// # Arguments
-    /// * `creator` - The pool creator (must match pool.creator)
-    /// * `pool_id` - The pool identifier
-    /// * `users` - Vector of addresses to add to the whitelist
-    ///
-    /// # Returns
-    /// Result indicating success or a PredifiError
-    ///
-    /// # Errors
-    /// * `Unauthorized` - If caller is not the pool creator
-    /// * `PoolNotFound` - If the pool doesn't exist
-    /// * `InvalidPoolState` - If pool is not private
-    /// * `InvalidData` - If users vector is empty or exceeds max batch size
-    pub fn batch_add_to_whitelist(
-        env: Env,
-        creator: Address,
-        pool_id: u64,
-        users: Vec<Address>,
-    ) -> Result<u32, PredifiError> {
-        const MAX_BATCH_SIZE: u32 = 100;
-
-        Self::require_not_paused(&env)?;
-        creator.require_auth();
-
-        // Validate batch size
-        let batch_size = users.len();
-        if batch_size == 0 {
-            return Err(PredifiError::InvalidData);
-        }
-        if batch_size > MAX_BATCH_SIZE {
-            return Err(PredifiError::InvalidData);
-        }
-
-        // Load and validate pool
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .ok_or(PredifiError::PoolNotFound)?;
-        Self::extend_persistent(&env, &pool_key);
-
-        // Authorization check
-        if pool.creator != creator {
-            return Err(PredifiError::Unauthorized);
-        }
-
-        // Pool must be private
-        if !pool.private {
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        // Batch process: add all users to whitelist
-        let mut added_count: u32 = 0;
-        let timestamp = env.ledger().timestamp();
-
-        for user in users.iter() {
-            let whitelist_key = DataKey::Whitelist(pool_id, user.clone());
-
-            // Only add if not already whitelisted
-            let already_whitelisted = env
-                .storage()
-                .persistent()
-                .get(&whitelist_key)
-                .unwrap_or(false);
-
-            if !already_whitelisted {
-                env.storage().persistent().set(&whitelist_key, &true);
-                Self::extend_persistent(&env, &whitelist_key);
-                added_count += 1;
-
-                // Emit event for each user
-                AddedToWhitelistEvent {
-                    pool_id,
-                    user,
-                    added_by: creator.clone(),
-                    timestamp,
-                }
-                .publish(&env);
-            }
-        }
-
-        Ok(added_count)
-    }
-
-    /// Batch remove multiple users from a private pool's whitelist.
-    ///
-    /// This optimized function reduces storage operations by processing
-    /// multiple users in a single transaction, improving gas efficiency.
-    ///
-    /// # Arguments
-    /// * `creator` - The pool creator (must match pool.creator)
-    /// * `pool_id` - The pool identifier
-    /// * `users` - Vector of addresses to remove from the whitelist
-    ///
-    /// # Returns
-    /// Result indicating success or a PredifiError
-    ///
-    /// # Errors
-    /// * `Unauthorized` - If caller is not the pool creator
-    /// * `PoolNotFound` - If the pool doesn't exist
-    /// * `InvalidPoolState` - If pool is not private
-    /// * `InvalidData` - If users vector is empty or exceeds max batch size
-    pub fn batch_remove_from_whitelist(
-        env: Env,
-        creator: Address,
-        pool_id: u64,
-        users: Vec<Address>,
-    ) -> Result<u32, PredifiError> {
-        const MAX_BATCH_SIZE: u32 = 100;
-
-        Self::require_not_paused(&env)?;
-        creator.require_auth();
-
-        // Validate batch size
-        let batch_size = users.len();
-        if batch_size == 0 {
-            return Err(PredifiError::InvalidData);
-        }
-        if batch_size > MAX_BATCH_SIZE {
-            return Err(PredifiError::InvalidData);
-        }
-
-        // Load and validate pool
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .ok_or(PredifiError::PoolNotFound)?;
-        Self::extend_persistent(&env, &pool_key);
-
-        // Authorization check
-        if pool.creator != creator {
-            return Err(PredifiError::Unauthorized);
-        }
-
-        // Pool must be private
-        if !pool.private {
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        // Batch process: remove all users from whitelist
-        let mut removed_count: u32 = 0;
-        let timestamp = env.ledger().timestamp();
-
-        for user in users.iter() {
-            let whitelist_key = DataKey::Whitelist(pool_id, user.clone());
-
-            // Only remove if currently whitelisted
-            if env.storage().persistent().has(&whitelist_key) {
-                env.storage().persistent().remove(&whitelist_key);
-                removed_count += 1;
-
-                // Emit event for each user
-                RemovedFromWhitelistEvent {
-                    pool_id,
-                    user,
-                    removed_by: creator.clone(),
-                    timestamp,
-                }
-                .publish(&env);
-            }
-        }
-
-        Ok(removed_count)
-    }
-
-    /// Batch check whitelist status for multiple users.
-    ///
-    /// This optimized function checks whitelist status for multiple users
-    /// in a single call, reducing RPC overhead for frontends.
-    ///
-    /// # Arguments
-    /// * `pool_id` - The pool identifier
-    /// * `users` - Vector of addresses to check
-    ///
-    /// # Returns
-    /// Vector of boolean values indicating whitelist status for each user
-    ///
-    /// # Errors
-    /// * `InvalidData` - If users vector is empty or exceeds max batch size
-    pub fn batch_check_whitelist(
-        env: Env,
-        pool_id: u64,
-        users: Vec<Address>,
-    ) -> Result<Vec<bool>, PredifiError> {
-        const MAX_BATCH_SIZE: u32 = 200;
-
-        // Validate batch size
-        let batch_size = users.len();
-        if batch_size == 0 {
-            return Err(PredifiError::InvalidData);
-        }
-        if batch_size > MAX_BATCH_SIZE {
-            return Err(PredifiError::InvalidData);
-        }
-
-        // Build result vector
-        let mut results = Vec::new(&env);
-
-        for user in users.iter() {
-            let whitelist_key = DataKey::Whitelist(pool_id, user);
-            let is_whitelisted = env
-                .storage()
-                .persistent()
-                .get(&whitelist_key)
-                .unwrap_or(false);
-
-            // Extend TTL for accessed keys
-            if env.storage().persistent().has(&whitelist_key) {
-                Self::extend_persistent(&env, &whitelist_key);
-            }
-
-            results.push_back(is_whitelisted);
-        }
-
-        Ok(results)
-    }
-
-    /// Return the number of unique participants in a pool.
-    ///
-    /// A participant is any address that has placed at least one prediction.
-    /// Subsequent top-ups by the same address do not increase the count.
-    ///
-    /// # Arguments
-    /// * `pool_id` - The unique identifier of the pool.
-    ///
-    /// # Returns
-    /// The number of unique participants as a `u32`.
-    pub fn get_pool_participants_count(env: Env, pool_id: u64) -> u32 {
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-        Self::extend_persistent(&env, &pool_key);
-        pool.participants_count
-    }
-
-    /// Get comprehensive stats for a pool.
-    ///
-    /// Gas notes: uses the in-pool `participants_count` and computes odds from the
-    /// already-loaded stakes vec via [`gas_opt::odds_from_stakes`].
-    pub fn get_pool_stats(env: Env, pool_id: u64) -> PoolStats {
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-        Self::extend_persistent(&env, &pool_key);
-
-        let stakes = Self::get_outcome_stakes(&env, pool_id, pool.options_count);
-        let current_odds = gas_opt::odds_from_stakes(&env, &stakes, pool.total_stake);
-
-        PoolStats {
-            pool_id,
-            total_stake: pool.total_stake,
-            stakes_per_outcome: stakes,
-            participants_count: pool.participants_count,
-            current_odds,
-        }
-    }
-
-    /// Initialize the oracle configuration. Only callable by Admin (role 0).
-    ///
-    /// # Errors
-    /// - `InvalidData`   – `max_price_age` is 0 (every feed would be immediately stale).
-    /// - `InvalidFeeBps` – `min_confidence_ratio` exceeds 10 000 bps (100 %).
-    pub fn init_oracle(
-        env: Env,
-        admin: Address,
-        pyth_contract: Address,
-        max_price_age: u64,
-        min_confidence_ratio: u32,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "init_oracle")?;
-
-        if max_price_age == 0 {
-            return Err(PredifiError::InvalidData);
-        }
-        if min_confidence_ratio > 10_000 {
-            return Err(PredifiError::InvalidFeeBps);
-        }
-
-        env.storage().persistent().set(
-            &DataKey::OracleConfig,
-            &(pyth_contract.clone(), max_price_age, min_confidence_ratio),
-        );
-        Self::extend_persistent(&env, &DataKey::OracleConfig);
-
-        OracleInitEvent {
-            admin,
-            pyth_contract,
-            max_price_age,
-            min_confidence_ratio,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Return the current oracle configuration, if initialised.
-    pub fn get_oracle_config(env: Env) -> Option<(Address, u64, u32)> {
-        env.storage()
-            .persistent()
-            .get::<DataKey, (Address, u64, u32)>(&DataKey::OracleConfig)
-    }
-
-    /// Set a price-based condition for automated pool resolution.
-    /// Only callable by Operator (role 1).
-    pub fn set_price_condition(
-        env: Env,
-        operator: Address,
-        pool_id: u64,
-        feed_pair: Symbol,
-        target_price: i128,
-        operator_type: u32,
-        tolerance_bps: u32,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        operator.require_auth();
-        Self::require_role(&env, &operator, 1)?; // Role Operator
-
-        if target_price <= 0 {
-            return Err(PredifiError::InvalidTargetPrice);
-        }
-
-        Self::validate_price_condition_match_params(operator_type, tolerance_bps)?;
-
-        let pool_key = DataKey::Pool(pool_id);
-        if !env.storage().persistent().has(&pool_key) {
-            return Err(PredifiError::PoolNotFound);
-        }
-
-        let condition_key = DataKey::PriceCondition(pool_id);
-        env.storage().persistent().set(
-            &condition_key,
-            &(
-                feed_pair.clone(),
-                target_price,
-                operator_type,
-                tolerance_bps,
-            ),
-        );
-        Self::extend_persistent(&env, &condition_key);
-
-        // Issue #1142: emit event for consistency with other setter functions.
-        PriceConditionSetEvent {
-            pool_id,
-            feed_pair,
-            target_price,
-            operator: operator_type,
-            tolerance_bps,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Update price feed data from an external oracle.
-    /// Only callable by authorized oracles.
-    pub fn update_price_feed(
-        env: Env,
-        oracle: Address,
-        feed_pair: Symbol,
-        price: i128,
-        confidence: i128,
-        timestamp: u64,
-        expires_at: u64,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        oracle.require_auth();
-
-        if !Self::is_oracle_whitelisted(&env, &oracle) {
-            return Err(PredifiError::Unauthorized);
-        }
-
-        // Security: reject prices with a future or current timestamp — the
-        // timestamp must be strictly in the past to prevent oracle manipulation
-        // via pre-dated or same-ledger price injections.
-        if timestamp >= env.ledger().timestamp() {
-            return Err(PredifiError::InvalidData);
-        }
-
-        // expires_at must be after the timestamp
-        if expires_at <= timestamp {
-            return Err(PredifiError::InvalidData);
-        }
-
-        let feed_key = DataKey::PriceFeed(feed_pair.clone());
-        env.storage()
-            .persistent()
-            .set(&feed_key, &(price, confidence, timestamp, expires_at));
-        Self::extend_persistent(&env, &feed_key);
-
-        // Track feed pair for cleanup
-        let mut list: Vec<Symbol> = env
-            .storage()
-            .persistent()
-            .get(&DataKey::PriceFeedList)
-            .unwrap_or_else(|| Vec::new(&env));
-        if !list.contains(feed_pair.clone()) {
-            list.push_back(feed_pair.clone());
-            env.storage()
-                .persistent()
-                .set(&DataKey::PriceFeedList, &list);
-        }
-
-        // Emit event so off-chain monitors and indexers can track price updates.
-        PriceFeedUpdatedEvent {
-            oracle,
-            feed_pair,
-            price,
-            confidence,
-            timestamp,
-            expires_at,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    /// Remove all expired price feeds from storage. Permissionless.
-    ///
-    /// Returns the number of feeds removed.
-    pub fn cleanup_expired_feeds(env: Env) -> u32 {
-        let current_time = env.ledger().timestamp();
-
-        let list: Vec<Symbol> = env
-            .storage()
-            .persistent()
-            .get(&DataKey::PriceFeedList)
-            .unwrap_or_else(|| Vec::new(&env));
-
-        let mut remaining: Vec<Symbol> = Vec::new(&env);
-        let mut removed: u32 = 0;
-
-        for i in 0..list.len() {
-            let pair = list.get(i).unwrap();
-            let expired = env
-                .storage()
-                .persistent()
-                .get::<DataKey, (i128, i128, u64, u64)>(&DataKey::PriceFeed(pair.clone()))
-                .map(|(_, _, _, expires_at)| expires_at < current_time)
-                .unwrap_or(true);
-
-            if expired {
-                env.storage().persistent().remove(&DataKey::PriceFeed(pair));
-                removed += 1;
-            } else {
-                remaining.push_back(pair);
-            }
-        }
-
-        env.storage()
-            .persistent()
-            .set(&DataKey::PriceFeedList, &remaining);
-
-        PriceFeedsCleanedEvent {
-            feeds_removed: removed,
-            timestamp: current_time,
-        }
-        .publish(&env);
-
-        removed
-    }
-
-    /// Load the current price and expiry timestamp for the configured feed.
-    ///
-    /// This intentionally preserves the previous missing-feed behavior:
-    /// callers panic with "Feed not found" when the feed has not been updated.
-    fn load_price_feed_for_resolution(env: &Env, feed_pair: Symbol) -> (i128, u64, u64) {
-        let (price, _confidence, timestamp, expires_at): (i128, i128, u64, u64) = env
-            .storage()
-            .persistent()
-            .get(&DataKey::PriceFeed(feed_pair))
-            .expect("Feed not found");
-
-        (price, timestamp, expires_at)
-    }
-
-    fn require_fresh_price_feed(
-        env: &Env,
-        timestamp: u64,
-        expires_at: u64,
-    ) -> Result<(), PredifiError> {
-        let current_time = env.ledger().timestamp();
-
-        // Check if price has expired
-        if current_time > expires_at {
-            return Err(PredifiError::PriceDataInvalid);
-        }
-
-        // Check if price is within the oracle's max_price_age limit
-        let (_pyth_contract, max_price_age, _min_confidence_ratio) = Self::get_oracle_config(env.clone())
-            .ok_or(PredifiError::OracleNotInitialized)?;
-        
-        if current_time > timestamp.saturating_add(max_price_age) {
-            return Err(PredifiError::PriceDataInvalid);
-        }
-
-        Ok(())
-    }
-
-    fn load_price_resolution_condition(
-        env: &Env,
-        pool_id: u64,
-    ) -> Result<(Symbol, i128, u32, u32), PredifiError> {
-        let condition_key = DataKey::PriceCondition(pool_id);
-        env.storage()
-            .persistent()
-            .get(&condition_key)
-            .ok_or(PredifiError::PriceConditionNotSet)
-    }
-
-    fn validate_price_condition_match_params(
-        comparison_op: u32,
-        tolerance_bps: u32,
-    ) -> Result<(), PredifiError> {
-        if comparison_op > 2 || tolerance_bps > MAX_TOLERANCE {
-            return Err(PredifiError::InvalidData);
-        }
-
-        Ok(())
-    }
-
-    fn price_tolerance_amount(
-        target_price: i128,
-        tolerance_bps: u32,
-    ) -> Result<i128, PredifiError> {
-        target_price
-            .checked_mul(tolerance_bps as i128)
-            .and_then(|amount| amount.checked_div(MAX_TOLERANCE as i128))
-            .ok_or(PredifiError::ArithmeticError)
-    }
-
-    /// Match a stored price condition in a fixed, explicitly bounded number of checks.
-    ///
-    /// Operators: 0=Equal, 1=Greater, 2=Less. Outcome: 0=No, 1=Yes.
-    fn price_resolution_outcome(
-        price: i128,
-        target_price: i128,
-        comparison_op: u32,
-        tolerance_bps: u32,
-    ) -> Result<u32, PredifiError> {
-        Self::validate_price_condition_match_params(comparison_op, tolerance_bps)?;
-
-        let tolerance_amount = Self::price_tolerance_amount(target_price, tolerance_bps)?;
-        let lower_bound = target_price
-            .checked_sub(tolerance_amount)
-            .ok_or(PredifiError::ArithmeticError)?;
-        let upper_bound = target_price
-            .checked_add(tolerance_amount)
-            .ok_or(PredifiError::ArithmeticError)?;
-
-        let mut steps = 0u32;
-        let condition_met = match comparison_op {
-            0 => {
-                steps += 2;
-                price >= lower_bound && price <= upper_bound
-            }
-            1 => {
-                steps += 1;
-                price > upper_bound
-            }
-            2 => {
-                steps += 1;
-                price < lower_bound
-            }
-            _ => return Err(PredifiError::InvalidData),
-        };
-
-        steps += 2;
-        if steps > MAX_PRICE_CONDITION_MATCH_STEPS {
-            return Err(PredifiError::RateLimitOrSuspiciousActivity);
-        }
-
-        Ok(if condition_met { 1 } else { 0 })
-    }
-
-    /// Load the pool and validate all non-outcome preconditions for price resolution.
-    fn load_resolvable_price_pool(
-        env: &Env,
-        pool_id: u64,
-    ) -> Result<(DataKey, Pool), PredifiError> {
-        let pool_key = DataKey::Pool(pool_id);
-        let pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-
-        Self::validate_pool_invariants(&pool);
-
-        if pool.state != MarketState::Active {
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        let current_time = env.ledger().timestamp();
-        let config = Self::get_config(env);
-
-        if current_time < pool.end_time.saturating_add(config.resolution_delay) {
-            return Err(PredifiError::ResolutionDelayNotMet);
-        }
-
-        Ok((pool_key, pool))
-    }
-
-    fn validate_price_resolution_outcome(
-        env: &Env,
-        pool_id: u64,
-        outcome: u32,
-        options_count: u32,
-    ) -> Result<(), PredifiError> {
-        if outcome >= options_count {
-            log!(
-                env,
-                "resolve_pool_from_price rejected: outcome is out of bounds",
-                pool_id,
-                outcome,
-                options_count
-            );
-            return Err(PredifiError::InvalidOutcome);
-        }
-
-        if outcome == UNRESOLVED_OUTCOME {
-            log!(
-                env,
-                "resolve_pool_from_price rejected: outcome cannot be sentinel value",
-                pool_id,
-                outcome
-            );
-            return Err(PredifiError::InvalidOutcome);
-        }
-
-        Ok(())
-    }
-
-    fn persist_price_resolution(
-        env: &Env,
-        pool_key: &DataKey,
-        pool_id: u64,
-        mut pool: Pool,
-        outcome: u32,
-    ) {
-        pool.state = MarketState::Resolved;
-        pool.outcome = outcome;
-        pool.fee_bps = Self::calculate_dynamic_fee(env, &pool);
-        pool.resolution_timestamp = Some(env.ledger().timestamp()); // Record resolution time
-
-        env.storage().persistent().set(pool_key, &pool);
-        Self::bump_ttl(env, pool_key);
-
-        PoolResolvedEvent {
-            pool_id,
-            operator: env.current_contract_address(),
-            outcome,
-        }
-        .publish(env);
-    }
-
-    /// Automatically resolve a pool based on its configured price condition.
-    /// Anyone can trigger this once the pool's end time and resolution delay have passed.
-    pub fn resolve_pool_from_price(env: Env, pool_id: u64) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-
-        let (feed_pair, target_price, comparison_op, tolerance_bps) =
-            Self::load_price_resolution_condition(&env, pool_id)?;
-        let (price, timestamp, expires_at) = Self::load_price_feed_for_resolution(&env, feed_pair);
-        Self::require_fresh_price_feed(&env, timestamp, expires_at)?;
-
-        let outcome =
-            Self::price_resolution_outcome(price, target_price, comparison_op, tolerance_bps)?;
-        let (pool_key, pool) = Self::load_resolvable_price_pool(&env, pool_id)?;
-        Self::validate_price_resolution_outcome(&env, pool_id, outcome, pool.options_count)?;
-        Self::persist_price_resolution(&env, &pool_key, pool_id, pool, outcome);
-
-        Ok(())
-    }
-    pub fn set_fee_tiers(
-        env: Env,
-        admin: Address,
-        tiers: Vec<FeeTier>,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "set_fee_tiers")?;
-
-        for i in 0..tiers.len() {
-            if let Some(tier) = tiers.get(i) {
-                if tier.fee_bps > 10_000 {
-                    return Err(PredifiError::InvalidFeeBps);
-                }
-                if i > 0 {
-                    if let Some(prev) = tiers.get(i - 1) {
-                        if tier.stake_threshold <= prev.stake_threshold {
-                            return Err(PredifiError::InvalidFeeBps);
-                        }
-                    }
-                }
-            }
-        }
-
-        env.storage().persistent().set(&DataKey::FeeTiers, &tiers);
-        Self::bump_ttl(&env, &DataKey::FeeTiers);
-
-        FeeTiersUpdateEvent {
-            admin,
-            tiers_count: tiers.len(),
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    pub fn get_fee_tiers(env: Env) -> Vec<FeeTier> {
-        env.storage()
-            .persistent()
-            .get(&DataKey::FeeTiers)
-            .unwrap_or_else(|| Vec::new(&env))
     }
 
     fn calculate_dynamic_fee(env: &Env, pool: &Pool) -> u32 {
@@ -6231,257 +2010,6 @@ impl PredifiContract {
         }
         applied_fee
     }
-
-    /// Emergency escape hatch: transfers any token balance held by this contract
-    /// to a destination address. Restricted to the admin role.
-    ///
-    /// Intended for use when the protocol or oracle has failed and funds must be
-    /// rescued. Emits an `EmergencyWithdraw` event for on-chain auditability.
-    pub fn emergency_withdraw(
-        env: Env,
-        admin: Address,
-        token: Address,
-        destination: Address,
-        amount: i128,
-    ) -> Result<(), PredifiError> {
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "emergency_withdraw")?;
-
-        // Validate token transfer before execution
-        Self::validate_token_transfer(
-            &env,
-            &token,
-            &env.current_contract_address(),
-            &destination,
-            amount,
-        )?;
-
-        let token_client = token::Client::new(&env, &token);
-
-        Self::enter_reentrancy_guard(&env);
-        token_client.transfer(&env.current_contract_address(), &destination, &amount);
-        Self::exit_reentrancy_guard(&env);
-
-        EmergencyWithdrawEvent {
-            admin,
-            token,
-            destination,
-            amount,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-}
-
-#[contractimpl]
-impl OracleCallback for PredifiContract {
-    fn oracle_resolve(
-        env: Env,
-        oracle: Address,
-        pool_id: u64,
-        outcome: u32,
-        proof: String,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        oracle.require_auth();
-
-        Self::require_oracle_role_for_resolution(&env, &oracle, pool_id)?;
-
-        let pool_key = DataKey::Pool(pool_id);
-        let mut pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-
-        Self::validate_pool_invariants(&pool);
-
-        // if pool.state != MarketState::Active {
-        //     return Err(PredifiError::InvalidPoolState);
-        // }
-        if !Self::is_pool_active(&pool) {
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        let current_time = env.ledger().timestamp();
-        let config = Self::get_config(&env);
-
-        if current_time < pool.end_time.saturating_add(config.resolution_delay) {
-            return Err(PredifiError::ResolutionDelayNotMet);
-        }
-
-        // Validate: outcome must be within the valid options range
-        if outcome >= pool.options_count {
-            soroban_sdk::panic_with_error!(&env, PredifiError::InvalidOutcome);
-        }
-
-        // Validate: outcome cannot be the sentinel value
-        if outcome == UNRESOLVED_OUTCOME {
-            soroban_sdk::panic_with_error!(&env, PredifiError::InvalidOutcome);
-        }
-
-        // --- Multi-oracle Voting Logic ---
-
-        let vote_key = DataKey::ResVote(pool_id, oracle.clone());
-        if env.storage().temporary().has(&vote_key) {
-            return Err(PredifiError::OracleAlreadyVoted);
-        }
-
-        // Record the oracle's vote in temporary storage
-        env.storage().temporary().set(&vote_key, &outcome);
-        Self::extend_temporary(&env, &vote_key);
-
-        // Increment total number of votes cast for this pool
-        let total_votes_key = DataKey::ResTotal(pool_id);
-        let total_votes: u32 = env.storage().temporary().get(&total_votes_key).unwrap_or(0);
-        let new_total_votes = total_votes + 1;
-        env.storage()
-            .temporary()
-            .set(&total_votes_key, &new_total_votes);
-        Self::extend_temporary(&env, &total_votes_key);
-
-        // Increment specific outcome vote count
-        let outcome_votes_key = DataKey::ResVoteCt(pool_id, outcome);
-        let outcome_votes: u32 = env
-            .storage()
-            .temporary()
-            .get(&outcome_votes_key)
-            .unwrap_or(0);
-        let new_outcome_votes = outcome_votes + 1;
-        env.storage()
-            .temporary()
-            .set(&outcome_votes_key, &new_outcome_votes);
-        Self::extend_temporary(&env, &outcome_votes_key);
-
-        // Detect conflicts: if there are ANY votes for a different outcome
-        if new_total_votes > new_outcome_votes {
-            // A conflict exists. Find at least one other voted outcome for the event.
-            for i in 0..pool.options_count {
-                if i == outcome {
-                    continue;
-                }
-                let other_key = DataKey::ResVoteCt(pool_id, i);
-                if env.storage().temporary().has(&other_key) {
-                    ResolutionConflictEvent {
-                        pool_id,
-                        oracle: oracle.clone(),
-                        outcome,
-                        existing_outcome: i,
-                    }
-                    .publish(&env);
-                    return Err(PredifiError::ResolutionConflict);
-                }
-            }
-        }
-
-        OracleResolvedEvent {
-            pool_id,
-            oracle: oracle.clone(),
-            outcome,
-            proof,
-        }
-        .publish(&env);
-
-        // Emit vote-cast event for oracle votes as well
-        ResolutionVoteCastEvent {
-            pool_id,
-            voter: oracle.clone(),
-            outcome,
-            vote_count: new_outcome_votes,
-            required_resolutions: pool.required_resolutions,
-        }
-        .publish(&env);
-
-        // Check if the required threshold has been met
-        if new_outcome_votes >= pool.required_resolutions {
-            pool.state = MarketState::Resolved;
-            pool.outcome = outcome;
-            pool.fee_bps = Self::calculate_dynamic_fee(&env, &pool);
-            pool.resolution_timestamp = Some(env.ledger().timestamp()); // Record resolution time
-
-            env.storage().persistent().set(&pool_key, &pool);
-            Self::bump_ttl(&env, &pool_key);
-
-            Self::extend_persistent(&env, &pool_key);
-            // Remove from global active index now that the pool is resolved.
-            Self::remove_from_active_index(&env, pool_id);
-
-            // Retrieve winning-outcome stake for the diagnostic event efficiently
-            let winning_stake = Self::get_outcome_stake(env.clone(), pool_id, outcome);
-
-            // Emit resolution events once threshold is met
-            PoolResolvedEvent {
-                pool_id,
-                operator: oracle,
-                outcome,
-            }
-            .publish(&env);
-
-            PoolResolvedDiagEvent {
-                pool_id,
-                outcome,
-                total_stake: pool.total_stake,
-                winning_stake,
-                timestamp: env.ledger().timestamp(),
-            }
-            .publish(&env);
-        }
-
-        Ok(())
-    }
-}
-
-#[contractimpl]
-impl PredifiContract {
-    /// Flag a pool as disputed. Only callable by a Moderator (role 2).
-    ///
-    /// Transitions the pool state from `Active` to `Disputed`, preventing
-    /// further participation or resolution until the dispute is handled.
-    ///
-    /// # Errors
-    /// - `Unauthorized` – caller does not hold the Moderator role.
-    /// - `InvalidPoolState` – pool is not currently `Active`.
-    pub fn flag_disputed_pool(
-        env: Env,
-        moderator: Address,
-        pool_id: u64,
-        reason: String,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        moderator.require_auth();
-        Self::require_role(&env, &moderator, 2)?;
-
-        let pool_key = DataKey::Pool(pool_id);
-        let mut pool: Pool = env
-            .storage()
-            .persistent()
-            .get(&pool_key)
-            .expect("Pool not found");
-        Self::extend_persistent(&env, &pool_key);
-
-        if pool.state != MarketState::Active {
-            return Err(PredifiError::InvalidPoolState);
-        }
-
-        pool.state = MarketState::Disputed;
-        env.storage().persistent().set(&pool_key, &pool);
-        Self::extend_persistent(&env, &pool_key);
-
-        env.storage()
-            .persistent()
-            .set(&DataKey::Disputed(pool_id), &());
-        Self::extend_persistent(&env, &DataKey::Disputed(pool_id));
-
-        PoolDisputedEvent {
-            pool_id,
-            moderator,
-            reason,
-        }
-        .publish(&env);
-
-        Ok(())
-    }
 }
 
 // ═══════════════════════════════════════════════════════════════════════════
@@ -6490,177 +2018,6 @@ impl PredifiContract {
 // Issue #1137 — Contract Metadata Getter
 // Issue #1142 — Event Emission Consistency (new events wired in below)
 // ═══════════════════════════════════════════════════════════════════════════
-
-#[contractimpl]
-impl PredifiContract {
-    // ── Issue #1125: Storage TTL renewal helper ───────────────────────────────
-
-    /// Renew (bump) the storage TTL for all persistent entries associated with
-    /// a pool, keeping them alive for another full `BUMP_AMOUNT` ledger period.
-    ///
-    /// This is a permissionless, read-only-effect helper that any party
-    /// (keeper bot, front-end, pool participant) can call at any time to
-    /// prevent pool data from expiring on-chain. It does **not** alter any
-    /// pool state — it is a pure TTL maintenance operation.
-    ///
-    /// Entries renewed:
-    /// - `Pool(pool_id)` — core pool struct
-    /// - `OutStakes(pool_id)` — batch outcome stakes (if present)
-    ///
-    /// # Arguments
-    /// * `pool_id` - The ID of the pool whose TTLs should be renewed.
-    ///
-    /// # Errors
-    /// * `PoolNotFound` — no pool exists for `pool_id`.
-    ///
-    /// # Events
-    /// Emits `StorageTtlRenewedEvent` so off-chain monitors can track
-    /// renewal activity and verify keeper health.
-    pub fn renew_storage_ttl(env: Env, pool_id: u64) -> Result<(), PredifiError> {
-        let pool_key = DataKey::Pool(pool_id);
-
-        // Verify the pool exists before attempting any TTL extension.
-        if !env.storage().persistent().has(&pool_key) {
-            return Err(PredifiError::PoolNotFound);
-        }
-
-        // Bump the pool struct entry.
-        Self::bump_ttl(&env, &pool_key);
-
-        // Bump batch outcome stakes entry if present.
-        let stakes_key = DataKey::OutStakes(pool_id);
-        if env.storage().persistent().has(&stakes_key) {
-            Self::extend_persistent(&env, &stakes_key);
-        }
-
-        StorageTtlRenewedEvent {
-            pool_id,
-            timestamp: env.ledger().timestamp(),
-        }
-        .publish(&env);
-
-        Ok(())
-    }
-
-    // ── Issue #1128: Referral volume threshold logic ──────────────────────────
-
-    /// Set the minimum referred volume (in base token units) that a referrer
-    /// must have accumulated in a pool before becoming eligible for a referral
-    /// reward on `claim_winnings`.
-    ///
-    /// A value of `0` disables the threshold — every referrer qualifies
-    /// regardless of volume (the original behaviour).
-    ///
-    /// This allows the protocol to filter out low-value or spam referrals that
-    /// might be created solely to extract small fee cuts without meaningful
-    /// traffic contribution.
-    ///
-    /// # Arguments
-    /// * `admin`      - Address with Admin role (0), must provide auth.
-    /// * `min_volume` - New threshold in base token units. Must be >= 0.
-    ///
-    /// # Errors
-    /// * `Unauthorized`   — caller does not hold the Admin role.
-    /// * `ContractPaused` — contract is currently paused.
-    pub fn set_referral_volume_threshold(
-        env: Env,
-        admin: Address,
-        min_volume: i128,
-    ) -> Result<(), PredifiError> {
-        Self::require_not_paused(&env)?;
-        admin.require_auth();
-        Self::require_admin_role(&env, &admin, "set_referral_volume_threshold")?;
-
-        if min_volume < 0 {
-            return Err(PredifiError::InvalidAmount);
-        }
-
-        env.storage()
-            .instance()
-            .set(&DataKey::ReferralMinVolumeBps, &min_volume);
-        Self::extend_instance(&env);
-
-        ReferralThresholdUpdatedEvent { admin, min_volume }.publish(&env);
-
-        Ok(())
-    }
-
-    /// Return the current referral volume threshold (in base token units).
-    ///
-    /// Returns `0` when no threshold has been configured (all referrers qualify).
-    pub fn get_referral_volume_threshold(env: Env) -> i128 {
-        env.storage()
-            .instance()
-            .get(&DataKey::ReferralMinVolumeBps)
-            .unwrap_or(0i128)
-    }
-
-    // ── Issue #1137: Contract metadata getter ─────────────────────────────────
-
-    /// Return a comprehensive metadata snapshot of the contract's current
-    /// configuration and operational state.
-    ///
-    /// This is designed as a single-call alternative to individually querying
-    /// `get_contract_info`, `get_fees`, `get_fee_tiers`, `get_oracle_config`,
-    /// `get_active_pools_count`, and `get_referral_volume_threshold`.
-    /// Front-ends and tooling can bootstrap their state from one RPC call.
-    ///
-    /// # Returns
-    /// A `ContractMetadata` struct containing all protocol parameters.
-    pub fn get_contract_metadata(env: Env) -> ContractMetadata {
-        let config = Self::get_config(&env);
-        let current_admin = Self::get_access_control_admin(&env, &config.access_control)
-            .unwrap_or_else(|_| env.current_contract_address());
-
-        let total_pools: u64 = env
-            .storage()
-            .instance()
-            .get(&DataKey::PoolIdCtr)
-            .unwrap_or(0u64);
-
-        let active_pools_count: u32 = env
-            .storage()
-            .persistent()
-            .get(&DataKey::ActivePoolCtr)
-            .unwrap_or(0u32);
-
-        let fee_tiers = Self::get_fee_tiers(env.clone());
-        let fee_tiers_count = fee_tiers.len();
-
-        let oracle_initialized = env.storage().persistent().has(&DataKey::OracleConfig);
-
-        let referral_min_volume: i128 = env
-            .storage()
-            .instance()
-            .get(&DataKey::ReferralMinVolumeBps)
-            .unwrap_or(0i128);
-
-        ContractMetadata {
-            version: env
-                .storage()
-                .instance()
-                .get(&DataKey::Version)
-                .unwrap_or(0u32),
-            version_string: Symbol::new(&env, "0_0_0"),
-            current_admin,
-            is_paused: Self::is_paused(&env),
-            total_pools,
-            active_pools_count,
-            fee_bps: config.fee_bps,
-            referral_cut_bps: Self::read_referral_cut_bps(&env),
-            referral_min_volume,
-            treasury: config.treasury,
-            access_control: config.access_control,
-            resolution_delay: config.resolution_delay,
-            min_pool_duration: config.min_pool_duration,
-            min_stake: config.min_stake,
-            max_predictions_per_user: config.max_predictions_per_user,
-            prediction_cooldown_seconds: config.prediction_cooldown_seconds,
-            fee_tiers_count,
-            oracle_initialized,
-        }
-    }
-}
 
 mod boundary_tests;
 mod edge_case_tests;
